% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_5p.R
\name{ab_5p}
\alias{ab_5p}
\title{Calculate mean antibody response using 5-parameter model}
\usage{
ab_5p(t, y0, y1, t1, alpha, shape)
}
\arguments{
\item{t}{time since last exposure}

\item{y0}{antibody response at exposure}

\item{y1}{antibody response at t1}

\item{t1}{time from last exposure to peak response}

\item{alpha}{decay power function coefficient}

\item{shape}{decay power function exponent
\link{numeric} scalar; uses r > 1 scale for shape}
}
\value{
a \link{numeric} \link{vector}
}
\description{
Calculate mean antibody response using 5-parameter model
}
\examples{
params <- typhoid_curves_nostrat_100[1, ]
ab_5p(
  t = units::as_units(50, "days"),
  y0 = params$y0,
  y1 = params$y1,
  t1 = params$t1 |> units::as_units("days"),
  alpha = params$alpha |> units::as_units("1/days"),
  shape = params$r)
}
\keyword{internal}
