% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.pop_data.R
\name{age_scatter}
\alias{age_scatter}
\title{Plot distribution of antibodies}
\usage{
age_scatter(object, strata = NULL)
}
\arguments{
\item{object}{A \code{pop_data} object (from \code{\link[=load_pop_data]{load_pop_data()}})}

\item{strata}{the name of a variable in \code{pop_data} to stratify by (or \code{NULL} for no stratification)}

\item{log}{whether to show antibody responses on logarithmic scale}

\item{...}{unused}

\item{type}{an option to choose type of chart:\code{density} or\code{age-scatter}}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot} object
}
\description{
\code{autoplot()} method for \code{pop_data} objects
}
\examples{
library(dplyr)
library(ggplot2)
library(ggfortify)

xs_data <- "https://osf.io/download//n6cp3/" \%>\%
  load_pop_data() \%>\%
  clean_pop_data()

xs_data \%>\% autoplot(strata = "Country", type = "density")
xs_data \%>\% autoplot(strata = "Country", type = "age-scatter")
}
