% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sr_params.R
\name{as_curve_params}
\alias{as_curve_params}
\title{Load antibody decay curve parameter}
\usage{
as_curve_params(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{as_curve_params()} was renamed to \code{\link[=as_sr_params]{as_sr_params()}} to create a more
consistent API.
}
\keyword{internal}
