% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.download_data.R
\name{autoplot.download_data}
\alias{autoplot.download_data}
\title{Plot download data with adaptive faceting}
\usage{
\method{autoplot}{download_data}(object, ...)
}
\arguments{
\item{object}{A \code{download_data} object from
\code{\link[=.get_download_data]{.get_download_data()}}.}

\item{...}{Not currently used.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Plot download data with adaptive faceting
}
\keyword{internal}
