% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.summary.estimate_scr_by.R
\name{autoplot.summary.estimate_scr_by}
\alias{autoplot.summary.estimate_scr_by}
\title{Plot method for \code{summary.estimate_scr_by} objects}
\usage{
\method{autoplot}{summary.estimate_scr_by}(
  object,
  xvar,
  fill_var,
  alpha = 0.7,
  CIs = FALSE,
  title = "Seroconversion Rate by Group",
  xlab = "Seroconversion rate per 1000 person-years",
  ylab = "Stratification Variable",
  fill_lab,
  fill_palette = NULL,
  ...
)
}
\arguments{
\item{object}{a \code{summary.seroincidence.by} object
(generated by applying the \code{summary()}
method to the output of \code{\link[=estimate_scr_by]{estimate_scr_by()}}).}

\item{xvar}{the name of a stratifying variable in \code{object} (e.g., "ageCat")}

\item{fill_var}{the name of the variable used for fill color (e.g., "Country")}

\item{alpha}{transparency for the bars (1 = no transparency, 0 = fully transparent)}

\item{CIs}{\link{logical}, if \code{TRUE}, add horizontal CI error bars}

\item{title}{a title for the final plot}

\item{xlab}{a label for the x axis of the final plot}

\item{ylab}{the name of a second stratifying variable in \code{object} (e.g., "catchment")}

\item{fill_lab}{fill label}

\item{fill_palette}{define the color palette for bar fill}

\item{...}{unused}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot method for \code{summary.estimate_scr_by} objects
}
\examples{

library(dplyr)
library(ggplot2)

xs_data <-
  sees_pop_data_pk_100

curve <-
  typhoid_curves_nostrat_100 \%>\%
  filter(antigen_iso \%in\% c("HlyE_IgA", "HlyE_IgG"))

noise <-
  example_noise_params_pk

est_catchment_age <- estimate_scr_by(
strata = c("catchment", "ageCat"),
pop_data = xs_data,
curve_params = curve,
curve_strata_varnames = NULL,
noise_params = noise,
noise_strata_varnames = "catchment",
antigen_isos = c("HlyE_IgG", "HlyE_IgA"),
num_cores = 2 # Allow for parallel processing to decrease run time
)

est_catchment_age_sum <- summary(est_catchment_age)

autoplot.summary.estimate_scr_by(
  est_catchment_age_sum,
  "ageCat",
  fill_var = "catchment"
)


}
