% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline}
\alias{baseline}
\title{Substitute baseline values}
\usage{
baseline(kab, yvec, blims, ...)
}
\arguments{
\item{kab}{\link{integer} indicating which row to read from \code{blims}}

\item{yvec}{a \link{numeric} \link{vector} of predicted biomarker values,
for one biomarker}

\item{blims}{range of possible baseline antibody levels}

\item{...}{unused}
}
\value{
an altered version of \code{yvec}
}
\description{
whenever y is below a cutoff (\code{blims[kab,2]}), substitute a random sample
from a baseline distribution
}
\keyword{internal}
