% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-help.R
\docType{data}
\name{campylobacterDelftParams1}
\alias{campylobacterDelftParams1}
\title{Campylobacter Delft Response Parameters Data for Model 1}
\format{A list of three dataframes:
\describe{
\item{\code{IgA}}{A dataframe containing 4000 rows with 7 parameters for IgA antibody.}
\item{\code{IgM}}{A dataframe containing 4000 rows with 7 parameters for IgM antibody.}
\item{\code{IgG}}{A dataframe containing 4000 rows with 7 parameters for IgG antibody.}
}}
\usage{
campylobacterDelftParams1
}
\description{
List of data frames of all longitudinal parameters. Each data frame contains
Monte Carlo samples for each antibody type.
}
\examples{

# Show first rows of every dataframe contained in campylobacterDelftParams1
lapply(campylobacterDelftParams1, head)

}
