% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-help.R
\docType{data}
\name{campylobacterDelftParams3}
\alias{campylobacterDelftParams3}
\title{Campylobacter Delft Response Parameters Data for Model 3}
\format{
A list of three dataframes:
\describe{
\item{\code{IgA}}{A dataframe containing 4000 rows with 7 parameters for IgA antibody.}
\item{\code{IgM}}{A dataframe containing 4000 rows with 7 parameters for IgM antibody.}
\item{\code{IgG}}{A dataframe containing 4000 rows with 7 parameters for IgG antibody.}
}
}
\usage{
campylobacterDelftParams3
}
\description{
List of data frames of all longitudinal parameters. Each data frame contains
Monte Carlo samples for each antibody type.
}
\examples{

# Show first rows of every dataframe contained in campylobacterDelftParams3
lapply(campylobacterDelftParams3, head)

}
