% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/campylobacterSerologyData.R
\docType{data}
\name{campylobacterSerologyData}
\alias{campylobacterSerologyData}
\title{Campylobacter Antibody Levels Data}
\format{A data frame with 258 observations on the following 3 variables:
\describe{
\item{\code{IgG}}{Measured IgG antibody level}
\item{\code{IgM}}{Measured IgM antibody level}
\item{\code{IgA}}{Measured IgA antibody level}
}}
\usage{
campylobacterSerologyData
}
\description{
Cross-sectional population sample of antibody levels data for Campylobacter.
}
\examples{
# show first rows of the data
head(campylobacterSerologyData)

# plot the data
plot(campylobacterSerologyData)
}

