% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-help.R
\docType{data}
\name{campylobacterSimLowData}
\alias{campylobacterSimLowData}
\alias{campylobacterSimMediumData}
\alias{campylobacterSimHighData}
\alias{pertussisSimLowData}
\alias{pertussisSimMediumData}
\alias{pertussisSimHighData}
\title{Simulated Cross-sectional Data}
\format{A data frame with 500 observations on the following 2 to 4 variables:
\describe{
\item{Age}{}
\item{IgG}{}
\item{IgM}{}
\item{IgA}{}
}}
\usage{
campylobacterSimLowData
}
\description{
Simulated cross-sectional population sample of antibody levels data for Campylobacter and
Pertussis for lambda 0.036/yr (low), 0.021/yr (medium) and 1.15/yr (high).
}
\examples{

# Show first rows of the data
head(campylobacterSimLowData)

# Summarize the data
summary(campylobacterSimLowData)

}
