% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parallel_cores.R
\name{check_parallel_cores}
\alias{check_parallel_cores}
\title{Check and Adjust Core Count Based on Hardware and Environment Limits}
\usage{
check_parallel_cores(num_cores, chk = "", verbose = FALSE)
}
\arguments{
\item{num_cores}{Integer. The initial number of requested cores.}

\item{chk}{Character. The value of an environment variable like \code{_R_CHECK_LIMIT_CORES_}. Defaults to \code{""}.}

\item{verbose}{Logical. Whether to report parallel setup information. Defaults to \code{FALSE}.}
}
\value{
An integer representing the safe number of cores to use.
}
\description{
Check and Adjust Core Count Based on Hardware and Environment Limits
}
\keyword{internal}
