% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pop_data.R
\name{check_pop_data}
\alias{check_pop_data}
\title{Check the formatting of a cross-sectional antibody survey dataset.}
\usage{
check_pop_data(pop_data, verbose = FALSE)
}
\arguments{
\item{pop_data}{dataset to check}

\item{verbose}{whether to print an "OK" message when all checks pass}
}
\value{
NULL (invisibly)
}
\description{
Check the formatting of a cross-sectional antibody survey dataset.
}
\examples{
library(magrittr)

xs_data <-
  serocalculator_example("example_pop_data.csv") |>
  read.csv() |>
  as_pop_data()

check_pop_data(xs_data, verbose = TRUE)

}
