% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_strata.R
\name{check_strata}
\alias{check_strata}
\title{Check a \code{pop_data} object for requested strata variables}
\usage{
check_strata(
  pop_data,
  strata,
  biomarker_names_var = get_biomarker_names_var(pop_data)
)
}
\arguments{
\item{pop_data}{a \code{pop_data} object}

\item{strata}{a \link{character} vector}

\item{biomarker_names_var}{name of column in \code{pop_data} indicating
biomarker type}
}
\value{
\link{NULL}, invisibly
}
\description{
Check a \code{pop_data} object for requested strata variables
}
\examples{
sees_pop_data_pk_100 |>
  check_strata(strata = c("ag", "catch", "Count")) |>
  try()
}
