% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_app.R
\name{curve_app}
\alias{curve_app}
\title{Launch the interactive Shiny app for the antigen-antibody kinetics model}
\usage{
curve_app()
}
\value{
\code{NULL} (invisibly)
}
\description{
Launch the interactive Shiny app for the antigen-antibody kinetics model
}
\examples{
\dontrun{
curve_app()
}
}
