% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_curve_params.R
\name{curve_converter}
\alias{curve_converter}
\title{Load antibody decay curve parameter}
\usage{
curve_converter(data, antigen_isos = NULL)
}
\arguments{
\item{data}{a \code{\link[=data.frame]{data.frame()}} or \link[tibble:tbl_df-class]{tibble::tbl_df}}

\item{antigen_isos}{\code{\link[=character]{character()}} vector of antigen isotypes to be used in analyses}
}
\value{
a \code{curve_data} object (a \link[tibble:tbl_df-class]{tibble::tbl_df} with extra attribute \code{antigen_isos})
}
\description{
Load antibody decay curve parameter
}
\examples{
curve_data <- load_curve_params("https://osf.io/download/rtw5k/")

print(curve_data)

}
