% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_summary.R
\name{.compute_kurtosis}
\alias{.compute_kurtosis}
\title{Compute Excess Kurtosis}
\usage{
.compute_kurtosis(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Excess kurtosis value
}
\description{
Internal function to compute excess kurtosis of a vector
}
\keyword{internal}
