% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_residuals.R
\name{.compute_predicted_values}
\alias{.compute_predicted_values}
\title{Compute Predicted Antibody Values}
\usage{
.compute_predicted_values(
  data,
  sr_params,
  noise_params,
  age_var,
  values_var,
  antigen_isos,
  verbose = FALSE
)
}
\arguments{
\item{data}{population data}

\item{sr_params}{serological response parameters}

\item{noise_params}{noise parameters}

\item{age_var}{name of age variable}

\item{values_var}{name of antibody values variable}

\item{antigen_isos}{antigens to compute predictions for}

\item{verbose}{logical}
}
\value{
A numeric vector of predicted values
}
\description{
Internal function to compute predicted antibody levels using the
fitted model parameters.
}
\keyword{internal}
