% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_summary.R
\name{.compute_pseudo_r2}
\alias{.compute_pseudo_r2}
\title{Compute Pseudo-R²}
\usage{
.compute_pseudo_r2(object, pop_data)
}
\arguments{
\item{object}{a seroincidence object}

\item{pop_data}{population data}
}
\value{
Pseudo-R² value
}
\description{
Internal function to compute pseudo-R² for the model
}
\keyword{internal}
