% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_summary.R
\name{.compute_skewness}
\alias{.compute_skewness}
\title{Compute Skewness}
\usage{
.compute_skewness(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Skewness value
}
\description{
Internal function to compute skewness of a vector
}
\keyword{internal}
