% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_residuals.R
\name{.estimate_residual_se}
\alias{.estimate_residual_se}
\title{Estimate Residual Standard Errors}
\usage{
.estimate_residual_se(noise_params, residuals_tbl, values_var)
}
\arguments{
\item{noise_params}{noise parameters}

\item{residuals_tbl}{residuals table}

\item{values_var}{name of values variable}
}
\value{
A numeric vector of standard error estimates
}
\description{
Internal function to estimate standard errors for residuals based on
noise parameters and observed values.
}
\keyword{internal}
