% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive_check.R
\name{.get_noise_limits}
\alias{.get_noise_limits}
\title{Extract Noise Limits from Noise Parameters}
\usage{
.get_noise_limits(noise_params)
}
\arguments{
\item{noise_params}{a noise_params tibble}
}
\value{
A matrix of noise limits suitable for sim_pop_data
}
\description{
Internal helper to extract noise limits matrix from noise_params tibble
}
\keyword{internal}
