% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optNll.R
\name{.optNll}
\alias{.optNll}
\title{Find the maximum likelihood estimate of the incidence rate parameter}
\usage{
.optNll(
  data = dataList$data,
  lnparams = dataList$lnparams,
  noise_params = dataList$noise_params,
  dataList = NULL,
  antigen_isos = unique(pull(data, antigen_iso)),
  lambda.start = exp(loglambda.start),
  loglambda.start = -6,
  log.lmin = loglambda.start - log(10),
  log.lmax = loglambda.start + log(10),
  stepmax = (log.lmax - log.lmin)/4,
  hessian = TRUE,
  ...
)
}
\arguments{
\item{data}{Data frame with cross-sectional serology data per antibody and age}

\item{lnparams}{List of data frames of all longitudinal parameters. Each data frame contains
Monte Carlo samples for each antibody type.}

\item{noise_params}{a \code{\link[=list]{list()}} (or \code{\link[=data.frame]{data.frame()}}, or \code{\link[=tibble]{tibble()}}) containing noise parameters}

\item{dataList}{Optional argument; as an alternative to passing in \code{data}, \code{lnparams}, and \code{noise_params} individually, you may create a list containing these three elements (with these names) and pass that in instead. This option may be useful for parallel processing across strata.}

\item{antigen_isos}{Character vector with one or more antibody names. Values must match \code{data}}

\item{lambda.start}{starting guess for incidence; only needed if \code{loglambda.start} is not used}

\item{loglambda.start}{starting guess for log(incidence); only needed if \code{lambda.start} is not used. A starting value for \code{log.lambda}. Value of -6 corresponds roughly to 1 day (\code{log(1/365.25)}), -4 corresponds roughly to 1 week (\code{log(7 / 365.25)}). Default = -6.}

\item{log.lmin}{used to determine default value of \code{stepmax}}

\item{log.lmax}{used to determine default value of \code{stepmax}}

\item{stepmax}{a positive scalar which gives the maximum allowable
    scaled step length.  \code{stepmax} is used to prevent steps which
    would cause the optimization function to overflow, to prevent the
    algorithm from leaving the area of interest in parameter space, or to
    detect divergence in the algorithm. \code{stepmax} would be chosen
    small enough to prevent the first two of these occurrences, but should
    be larger than any anticipated reasonable step.}

\item{hessian}{if \code{TRUE}, the hessian of \code{f}
    at the minimum is returned.}

\item{...}{
  Arguments passed on to \code{\link[=.nll]{.nll}}, \code{\link[stats:nlm]{stats::nlm}}
  \describe{
    \item{\code{antibodies}}{Character vector with one or more antibody names. Values must match \code{data}.}
    \item{\code{log.lambda}}{natural logarithm of incidence parameter, in log(years). Value of -6 corresponds roughly to 1 day (log(1/365.25)), -4 corresponds roughly to 1 week (log(7 / 365.25)).}
    \item{\code{f}}{the function to be minimized, returning a single numeric
    value.  This should be a function with first argument a vector of
    the length of \code{p} followed by any other arguments specified by
    the \code{\dots} argument.

    If the function value has an attribute called \code{gradient} or
    both \code{gradient} and \code{hessian} attributes, these will be
    used in the calculation of updated parameter values.  Otherwise,
    numerical derivatives are used. \code{\link[stats]{deriv}} returns a
    function with suitable \code{gradient} attribute and optionally a
    \code{hessian} attribute.}
    \item{\code{p}}{starting parameter values for the minimization.}
    \item{\code{typsize}}{an estimate of the size of each parameter
    at the minimum.}
    \item{\code{fscale}}{an estimate of the size of \code{f} at the minimum.}
    \item{\code{print.level}}{this argument determines the level of printing
    which is done during the minimization process.  The default
    value of \code{0} means that no printing occurs, a value of \code{1}
    means that initial and final details are printed and a value
    of 2 means that full tracing information is printed.}
    \item{\code{ndigit}}{the number of significant digits in the function \code{f}.}
    \item{\code{gradtol}}{a positive scalar giving the tolerance at which the
    scaled gradient is considered close enough to zero to
    terminate the algorithm.  The scaled gradient is a
    measure of the relative change in \code{f} in each direction
    \code{p[i]} divided by the relative change in \code{p[i]}.}
    \item{\code{steptol}}{A positive scalar providing the minimum allowable
    relative step length.}
    \item{\code{iterlim}}{a positive integer specifying the maximum number of
    iterations to be performed before the program is terminated.}
    \item{\code{check.analyticals}}{a logical scalar specifying whether the
    analytic gradients and Hessians, if they are supplied, should be
    checked against numerical derivatives at the initial parameter
    values. This can help detect incorrectly formulated gradients or
    Hessians.}
  }}
}
\value{
a \code{\link[stats:nlm]{stats::nlm()}} fit object
}
\description{
Find the maximum likelihood estimate of the incidence rate parameter
}
