% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclass.R
\name{.subclass}
\alias{.subclass}
\title{Prepend a subclass to an object's class vector}
\usage{
.subclass(x, subclass)
}
\arguments{
\item{x}{An object.}

\item{subclass}{Character string to prepend to
the class vector.}
}
\value{
\code{x} with \code{subclass} prepended to its class.
}
\description{
Prepend a subclass to an object's class vector
}
\keyword{internal}
