% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_seroincidence_by.R
\name{est.incidence.by}
\alias{est.incidence.by}
\title{Estimate Seroincidence}
\usage{
est.incidence.by(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{est.incidence.by()} was renamed to \code{\link[=est_seroincidence_by]{est_seroincidence_by()}} to create a more
consistent API.
}
\keyword{internal}
