% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_seroincidence.R
\name{est.incidence}
\alias{est.incidence}
\title{Estimate Seroincidence}
\usage{
est.incidence(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{est.incidence()} was renamed to \code{\link[=est_seroincidence]{est_seroincidence()}} to create a more
consistent API.
}
\keyword{internal}
