% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSeroincidence.R
\name{estimateSeroincidence}
\alias{estimateSeroincidence}
\title{Estimate Seroincidence}
\usage{
estimateSeroincidence(
  data,
  antibodies,
  strata = "",
  lnparams,
  noise_params,
  numCores = 1L,
  ...
)
}
\arguments{
\item{data}{Data frame with cross-sectional serology data per antibody and age, and additional
columns to identify possible \code{strata}.}

\item{antibodies}{Character vector with one or more antibody names. Values must match \code{data}.}

\item{strata}{Character vector of stratum-defining variables. Values must match with \code{data}. Default = "".}

\item{lnparams}{List of data frames of all longitudinal parameters. Each data frame contains
Monte Carlo samples for each antibody type.}

\item{noise_params}{a \code{\link[=list]{list()}} (or \code{\link[=data.frame]{data.frame()}}, or \code{\link[=tibble]{tibble()}}) containing noise parameters}

\item{numCores}{Number of processor cores to use for calculations when computing by strata. If
set to more than 1 and package \pkg{parallel} is available, then the computations are
executed in parallel. Default = 1L.}

\item{...}{
  Arguments passed on to \code{\link[=.optNll]{.optNll}}
  \describe{
    \item{\code{lambda.start}}{starting guess for incidence; only needed if \code{loglambda.start} is not used}
    \item{\code{loglambda.start}}{starting guess for log(incidence); only needed if \code{lambda.start} is not used. A starting value for \code{log(lambda)}. Value of -6 corresponds roughly to 1 day (\code{log(1/365.25)}), -4 corresponds roughly to 1 week (\code{log(7 / 365.25)}). Default = -6.}
    \item{\code{log.lmin}}{used to determine default value of \code{stepmax}}
    \item{\code{log.lmax}}{used to determine default value of \code{stepmax}}
    \item{\code{hessian}}{if \code{TRUE}, the hessian of \code{f}
    at the minimum is returned.}
    \item{\code{stepmax}}{a positive scalar which gives the maximum allowable
    scaled step length.  \code{stepmax} is used to prevent steps which
    would cause the optimization function to overflow, to prevent the
    algorithm from leaving the area of interest in parameter space, or to
    detect divergence in the algorithm. \code{stepmax} would be chosen
    small enough to prevent the first two of these occurrences, but should
    be larger than any anticipated reasonable step.}
  }}
}
\value{
A set of lambda estimates for each strata.
}
\description{
Function to estimate seroincidences based on cross-section serology data and longitudinal
response model.
}
\examples{

\dontrun{
estimateSeroincidence(data = csData,
                      antibodies = c("IgG", "IgM", "IgA"),
                      strata = "",
                      params = campylobacterDelftParams4,
                      censorLimits = cutOffs,
                      par0 = baseLn,
                      start = -4)

estimateSeroincidence(data = csData,
                      antibodies = c("IgG", "IgM", "IgA"),
                      strata = "",
                      params = campylobacterDelftParams4,
                      censorLimits = cutOffs,
                      par0 = baseLn,
                      start = -4,
                      numCores = parallel::detectCores())
}

}
