% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_snapshot_data.R
\name{expect_snapshot_data}
\alias{expect_snapshot_data}
\title{Snapshot testing for \link{data.frame}s}
\usage{
expect_snapshot_data(x, name, digits = 6)
}
\arguments{
\item{x}{a \link{data.frame} to snapshot}

\item{name}{\link{character} snapshot name}

\item{digits}{\link{integer} passed to \code{\link[=signif]{signif()}} for numeric variables}
}
\value{
\link{NULL} (from \code{\link[testthat:expect_snapshot_file]{testthat::expect_snapshot_file()}})
}
\description{
copied from \url{https://github.com/bcgov/ssdtools}
with permission (\url{https://github.com/bcgov/ssdtools/issues/379})
}
\examples{
\dontrun{
expect_snapshot_data(iris, name = "iris")
}
}
\keyword{internal}
