% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_dev.r
\name{fdev}
\alias{fdev}
\title{Calculate negative log-likelihood (deviance)}
\usage{
fdev(lambda, csdata, lnpars, cond)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{fdev()} was renamed to \code{\link[=f_dev]{f_dev()}} to create a more
consistent API.
}
\keyword{internal}
