% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sim_table.R
\name{generate_sim_table}
\alias{generate_sim_table}
\title{Generate table of simulation results}
\usage{
generate_sim_table(
  results_list,
  sample_size = attr(results_list, "sample_size")
)
}
\arguments{
\item{results_list}{output from \code{\link[=simulate_seroincidence]{simulate_seroincidence()}}}

\item{sample_size}{sample size of simulated data sets}
}
\value{
a \link[tibble:tbl_df-class]{tibble::tbl_df}
}
\description{
Generate table of simulation results
}
