% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_antigen_isos.R
\name{get_antigen_isos}
\alias{get_antigen_isos}
\title{Extract antigen-isotype combinations list from an object}
\usage{
get_antigen_isos(object)
}
\arguments{
\item{object}{any R object}
}
\value{
the antigen isotype combinations associated with that object, if any exist
}
\description{
Extract antigen-isotype combinations list from an object
}
\examples{

"https://osf.io/download//n6cp3/" |>
 load_pop_data() |>
 clean_pop_data() |>
 get_antigen_isos()
}
