% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_attributes.R
\name{get_biomarker_levels}
\alias{get_biomarker_levels}
\title{Extract biomarker levels}
\usage{
get_biomarker_levels(object, ...)
}
\arguments{
\item{object}{a \code{pop_data} object}

\item{...}{unused}
}
\value{
the biomarker levels in \code{object}
}
\description{
Extract biomarker levels
}
\examples{
sees_pop_data_100 |> get_biomarker_levels()
}
