% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biomarker_names.R
\name{get_biomarker_names}
\alias{get_biomarker_names}
\title{Extract biomarker names column}
\usage{
get_biomarker_names(object, ...)
}
\arguments{
\item{object}{a long-format dataset (\code{pop_data}, \code{curve_params}, etc)}

\item{...}{unused}
}
\value{
a \link{character} or \link{factor} \link{vector} of biomarker names
}
\description{
Extract biomarker names column
}
\examples{
sees_pop_data_100 |> get_biomarker_names() |> head()
}
\keyword{internal}
