% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biomarker_names_var.R
\name{get_biomarker_names_var}
\alias{get_biomarker_names_var}
\title{Get biomarker variable name}
\usage{
get_biomarker_names_var(object, ...)
}
\arguments{
\item{object}{a \code{pop_data} object}

\item{...}{unused}
}
\value{
a \link{character} string identifying the biomarker names column in \code{object}
}
\description{
Get biomarker variable name
}
\examples{
sees_pop_data_100 |> get_biomarker_names_var()
}
