% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_attributes.R
\name{get_values}
\alias{get_values}
\title{Get antibody measurement values}
\usage{
get_values(object, ...)
}
\arguments{
\item{object}{a \code{pop_data} object}

\item{...}{unused}
}
\value{
a \link{numeric} \link{vector} of antibody measurement values
}
\description{
Get antibody measurement values
}
\examples{
sees_pop_data_100 |> get_values()
}
