% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_attributes.R
\name{get_values_var}
\alias{get_values_var}
\title{Extract antibody measurement values}
\usage{
get_values_var(object, ...)
}
\arguments{
\item{object}{a \code{pop_data} object}

\item{...}{unused}
}
\value{
the name of the column in \code{object} specified as containing
antibody abundance measurements
}
\description{
Extract antibody measurement values
}
\examples{
sees_pop_data_100 |> get_values_var()
}
