% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.decay.curves.by.R
\name{graph.decay.curves.by}
\alias{graph.decay.curves.by}
\title{graph antibody decay curves by antigen isotype}
\usage{
graph.decay.curves.by(
  curve_params,
  antigen_isos = unique(curve_params$antigen_iso),
  ncol = min(3, length(antigen_isos)),
  ...
)
}
\arguments{
\item{curve_params}{\code{\link[=data.frame]{data.frame()}} of curve parameters (one or more MCMC samples)}

\item{antigen_isos}{antigen isotypes to analyze (can be used to subset \code{curve_params})}

\item{ncol}{how many columns of subfigures to use in panel plot}

\item{...}{
  Arguments passed on to \code{\link[=graph.decay.curves]{graph.decay.curves}}
  \describe{
    \item{\code{verbose}}{verbose output}
    \item{\code{xlim}}{range of x values to graph}
    \item{\code{n_curves}}{how many curves to plot (see details).}
    \item{\code{rows_to_graph}}{which rows of \code{curve_params} to plot (overrides \code{n_curves}).}
    \item{\code{alpha}}{(passed to \code{\link[ggplot2:geom_function]{ggplot2::geom_function()}}) how transparent the curves should be:
\itemize{
\item 0 = fully transparent (invisible)
\item 1 = fully opaque (no transparency)
}}
  }}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
graph antibody decay curves by antigen isotype
}
\details{
\subsection{\code{rows_to_graph}}{

Note that if you directly specify \code{rows_to_graph} when calling this function, the row numbers are enumerated separately for each antigen isotype; in other words, for the purposes of this argument, row numbers start over at 1 for each antigen isotype. There is currently no way to specify different row numbers for different antigen isotypes; if you want to do that, you will could call \code{\link[=graph.decay.curves]{graph.decay.curves()}} directly for each antigen isotype and combine the resulting panels yourself. Or you could subset \code{curve_params} manually, before passing it to this function, and set the \code{n_curves} argument to \code{Inf}.
}
}
