% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.decay.curves.R
\name{graph.decay.curves}
\alias{graph.decay.curves}
\title{Graph an antibody decay curve model}
\usage{
graph.decay.curves(
  curve_params,
  verbose = FALSE,
  alpha = 0.4,
  n_curves = 100,
  rows_to_graph = sample.int(n = nrow(curve_params), size = min(n_curves,
    nrow(curve_params)), replace = FALSE),
  xlim = c(10^-1, 10^3.1),
  ...
)
}
\arguments{
\item{curve_params}{\code{\link[=data.frame]{data.frame()}} of curve parameters (one or more MCMC samples)}

\item{verbose}{verbose output}

\item{alpha}{(passed to \code{\link[ggplot2:geom_function]{ggplot2::geom_function()}}) how transparent the curves should be:
\itemize{
\item 0 = fully transparent (invisible)
\item 1 = fully opaque (no transparency)
}}

\item{n_curves}{how many curves to plot (see details).}

\item{rows_to_graph}{which rows of \code{curve_params} to plot (overrides \code{n_curves}).}

\item{xlim}{range of x values to graph}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Graph an antibody decay curve model
}
\details{
\subsection{\code{n_curves} and \code{rows_to_graph}}{

In most cases, \code{curve_params} will contain too many rows of MCMC samples for all of these samples to be plotted at once.
\itemize{
\item Setting the  \code{n_curves} argument to a value smaller than the number of rows in \code{curve_params} will cause this function to randomly select (without replacement) a subset of \code{n_curves} rows to graph.
\item Setting \code{n_curves} larger than the number of rows in ` will result all curves being plotted.
\item If the user directly specifies the \code{rows_to_graph} argument, then \code{n_curves} has no effect.
}
}
}
\examples{
\dontrun{
curve_params = readRDS(url("https://osf.io/download/rtw5k/"))
plot1 = graph.curve.params(curve_params)
print(plot1)
}
}
