% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids.R
\name{ids}
\alias{ids}
\title{Get person IDs}
\usage{
ids(object, ...)
}
\arguments{
\item{object}{a \code{\link[=data.frame]{data.frame()}}}

\item{...}{unused}
}
\value{
a \link{character} \link{vector}
}
\description{
Get person IDs
}
\examples{
ids(sees_pop_data_pk_100)
}
\keyword{internal}
