% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids.R
\name{ids_varname}
\alias{ids_varname}
\title{Get person ID variable name}
\usage{
ids_varname(object, ...)
}
\arguments{
\item{object}{a \link{data.frame}}

\item{...}{unused}
}
\value{
a \link{character} string containing the person ID column,
as recorded in the metadata of \code{object}
}
\description{
Get person ID variable name
}
\examples{
ids_varname(sees_pop_data_pk_100)
}
\keyword{internal}
