% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estIncidence.R
\name{incidence.age}
\alias{incidence.age}
\title{Age specific seroincidence function
add some details here}
\usage{
incidence.age(
  dpop,
  dmcmc,
  cond,
  c.age,
  antigen_isos,
  noise_params = filter(cond, Country == "MGH"),
  start = 0.1,
  iterlim = 100
)
}
\arguments{
\item{dpop}{cross-sectional population data}

\item{dmcmc}{mcmc samples from distribution of longitudinal decay curve parameters}

\item{cond}{conditional noise parameters}

\item{c.age}{age category}

\item{antigen_isos}{antigen-isotype(s) (a \code{\link[=character]{character()}} vector of one or more antigen names)}

\item{noise_params}{a \code{\link[=data.frame]{data.frame()}} containing columns \code{nu}, etc.}

\item{start}{starting value for liklihood}

\item{iterlim}{a positive integer specifying the maximum number of iterations to be performed before the program is terminated.}
}
\value{
A \code{\link[=data.frame]{data.frame()}} containing the following:
\itemize{
\item \code{startingval}: the starting guess for incidence rate
\item \code{ageCat}: the age category we are analyzing
\item \code{incidence.rate}: the estimated incidence rate, per person year
\item \code{CI.lwr}: lower limit of confidence interval for incidence rate
\item \code{CI.upr}: upper limit of confidence interval for incidence rate
\item \code{coverage}: coverage probability
\item \code{neg.llik}: negative log-likelihood
\item \code{iterations}: the number of iterations used
}
}
\description{
Age specific seroincidence function
add some details here
}
