% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldpar.R
\name{ldpar}
\alias{ldpar}
\title{extract a row from longitudinal parameter set}
\usage{
ldpar(age, antigen_isos, nmc, npar, ...)
}
\arguments{
\item{age}{age at infection}

\item{antigen_isos}{antigen isotypes}

\item{nmc}{mcmc sample to use}

\item{npar}{number of parameters}

\item{...}{passed to \code{simpar()}}
}
\value{
an array of 7 parameters and initial conditions:
c(y0,b0,mu0,mu1,c1,alpha,shape)
}
\description{
take a random sample from longitudinal parameter set
given age at infection, for a list of antibodies
}
\keyword{internal}
