% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llik.R
\name{llik}
\alias{llik}
\title{Calculate log-likelihood}
\usage{
llik(
  lambda,
  pop_data,
  antigen_isos,
  curve_params,
  noise_params,
  verbose = FALSE,
  ...
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{llik()} was renamed to \code{\link[=log_likelihood]{log_likelihood()}} to create a more
consistent API.
}
\keyword{internal}
