% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{llik_vec}
\alias{llik_vec}
\title{Calculate log-likelihood (vectorized)}
\usage{
llik_vec(
  log.lambda = log(lambda),
  lambda = exp(log.lambda),
  data,
  antigen_isos,
  curve_params,
  noise_params,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{log.lambda}{natural logarithm of incidence parameter, in log(events per person-year).
\itemize{
\item Value of -6 corresponds roughly to 1 day (\code{log(1/365.25)})
\item Value of -4 corresponds roughly to 1 week (\code{log(7 / 365.25)}).
}}

\item{lambda}{incidence parameter, in events per person-year}

\item{data}{Data frame with cross-sectional serology data per antibody and age, and additional columns}

\item{antigen_isos}{Character vector with one or more antibody names. Values must match \code{data}.}

\item{curve_params}{List of data frames of all longitudinal parameters. Each data frame contains
Monte Carlo samples for each antibody type.}

\item{noise_params}{a \code{\link[=list]{list()}} (or \code{\link[=data.frame]{data.frame()}}, or \code{\link[=tibble]{tibble()}}) containing noise parameters}

\item{verbose}{logical: if TRUE, print verbose log information to console}

\item{...}{additional arguments passed to other functions (not currently used).}
}
\value{
the log-likelihood of the data with the current parameter values
}
\description{
Calculate log-likelihood (vectorized)
}
\details{
Same as \code{\link[=llik]{llik()}}, except vectorized for the \code{log.lambda} argument.
}
