% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sr_params.R
\name{load_curve_params}
\alias{load_curve_params}
\title{Load antibody decay curve parameter samples}
\usage{
load_curve_params(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{load_curve_params()} was renamed to \code{\link[=load_sr_params]{load_sr_params()}} to create a more
consistent API.
}
\keyword{internal}
