% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.cs-utils.r
\name{mk_baseline}
\alias{mk_baseline}
\title{generate random sample from baseline distribution}
\usage{
mk_baseline(kab, n = 1, blims, ...)
}
\arguments{
\item{kab}{index for which row of antibody baseline limits to read from \code{blims}}

\item{n}{number of observations}

\item{blims}{range of possible baseline antibody levels}

\item{...}{not currently used}
}
\value{
a \code{\link[=numeric]{numeric()}} vector
}
\description{
generate random sample from baseline distribution
}
