% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.cs-utils.r
\name{mkbaseline}
\alias{mkbaseline}
\title{generate random sample from baseline distribution}
\usage{
mkbaseline(kab, n = 1, blims, ...)
}
\arguments{
\item{kab}{index for which row of antibody baseline limits to read from \code{blims}}

\item{n}{number of observations}

\item{blims}{range of possible baseline antibody levels}

\item{...}{not currently used}
}
\value{
a \code{\link[=numeric]{numeric()}} vector
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{mkbaseline()} was renamed to \code{mk_baseline()} to create a more
consistent API.
}
\keyword{internal}
