% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen_decay_curve.R
\name{pathogen_decay_curve}
\alias{pathogen_decay_curve}
\title{Pathogen decay curve}
\usage{
pathogen_decay_curve(
  t,
  y0 = 0.74916052,
  b0 = 1,
  mu_b = 0.18432798,
  mu_y = 0.36853621,
  gamma = 0.0013040664
)
}
\arguments{
\item{t}{time since infection}

\item{y0}{initial antibody concentration}

\item{b0}{initial bacteria concentration}

\item{mu_b}{pathogen reproduction rate}

\item{mu_y}{antibody reproduction rate}

\item{gamma}{bacteria destruction rate per antibody}
}
\value{
a \code{\link[=numeric]{numeric()}}
}
\description{
Pathogen decay curve
}
\examples{
library(ggplot2)
ggplot() + geom_function(
  fun = pathogen_decay_curve,
  args = list(
    y0 = 0.74916052,
    b0 = 1,
    mu_b = 0.18432798,
    mu_y = 0.36853621,
    gamma = 0.0013040664
  )
) + xlim(0, 100)
}
