% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-help.R
\docType{data}
\name{pertussisIgGPTParams2}
\alias{pertussisIgGPTParams2}
\title{Pertussis IgG-PT Response Parameters Data for Model 2}
\format{A dataframe \code{IgG} containing 3000 rows with 7 parameters for IgG antibody.}
\usage{
pertussisIgGPTParams2
}
\description{
List of data frames of all longitudinal parameters. Each data frame contains
Monte Carlo samples for each antibody type.
}
\examples{

# Show first rows of every dataframe contained in pertussisIgGPTParams2
lapply(pertussisIgGPTParams2, head)

}
