% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.seroincidence.ests.R
\name{plot.seroincidence.by}
\alias{plot.seroincidence.by}
\title{Plot \code{seroincidence.by} log-likelihoods}
\usage{
\method{plot}{seroincidence.by}(x, ...)
}
\arguments{
\item{x}{a '"seroincidence.by"' object (from \code{\link[=est.incidence.by]{est.incidence.by()}})}

\item{...}{
  Arguments passed on to \code{\link[=plot.seroincidence]{plot.seroincidence}}
  \describe{
    \item{\code{log_x}}{should the x-axis be on a logarithmic scale (\code{TRUE}) or linear scale (\code{FALSE}, default)?}
  }}
}
\value{
aan object of class \code{"ggarrange"}, which is a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} or a \code{\link[=list]{list()}} of \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}s.
}
\description{
Plot \code{seroincidence.by} log-likelihoods
}
