% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.seroincidence.ests.R
\name{plot.seroincidence.ests}
\alias{plot.seroincidence.ests}
\title{Plot \code{seroincidence.ests} log-likelihoods}
\usage{
\method{plot}{seroincidence.ests}(object)
}
\arguments{
\item{object}{a '"seroincidence.ests"' object (from \code{\link[=est.incidence.by]{est.incidence.by()}})}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot \code{seroincidence.ests} log-likelihoods
}
