% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.seroincidence.est.R
\name{plot.seroincidence}
\alias{plot.seroincidence}
\title{Plot the log-likelihood curve for the incidence rate estimate}
\usage{
\method{plot}{seroincidence}(x, ...)
}
\arguments{
\item{x}{\code{seroincidence} object (from \code{\link[=est.incidence]{est.incidence()}})}

\item{...}{unused}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
\description{
Plot the log-likelihood curve for the incidence rate estimate
}
