% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.summary.seroincidence.ests.R
\name{plot.summary.seroincidence.ests}
\alias{plot.summary.seroincidence.ests}
\title{Plot method for \code{summary.seroincidence.ests} objects}
\usage{
\method{plot}{summary.seroincidence.ests}(x, xvar, alpha = 0.9, ...)
}
\arguments{
\item{x}{a \code{summary.seroincidence.ests} object (generated by applying the \code{summary()} method to the output of \code{\link[=est.incidence.by]{est.incidence.by()}}).}

\item{xvar}{the name of a stratifying variable in \code{object}}

\item{alpha}{transparency for the points in the graph (1 = no transparency, 0 = fully transparent)}

\item{...}{unused}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot method for \code{summary.seroincidence.ests} objects
}
