% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curve_params_one_ab.R
\name{plot_curve_params_one_ab}
\alias{plot_curve_params_one_ab}
\title{Graph an antibody decay curve model}
\usage{
plot_curve_params_one_ab(
  object,
  verbose = FALSE,
  alpha = 0.4,
  n_curves = 100,
  n_points = 1000,
  log_x = FALSE,
  log_y = TRUE,
  iters_to_graph = seq_len(min(n_curves, nrow(object))),
  xlim = c(10^-1, 10^3.1),
  ...
)
}
\arguments{
\item{object}{a \code{\link[=data.frame]{data.frame()}} of curve parameters (one or more MCMC samples)}

\item{verbose}{verbose output}

\item{alpha}{(passed to \code{\link[ggplot2:geom_function]{ggplot2::geom_function()}})
how transparent the curves should be:
\itemize{
\item 0 = fully transparent (invisible)
\item 1 = fully opaque
}}

\item{n_curves}{how many curves to plot (see details).}

\item{n_points}{Number of points to interpolate along the x axis
(passed to \code{\link[ggplot2:geom_function]{ggplot2::geom_function()}})}

\item{log_x}{should the x-axis be on a logarithmic scale (\code{TRUE})
or linear scale (\code{FALSE}, default)?}

\item{log_y}{should the Y-axis be on a logarithmic scale
(default, \code{TRUE}) or linear scale (\code{FALSE})?}

\item{iters_to_graph}{which MCMC iterations in \code{curve_params} to plot
(overrides \code{n_curves}).}

\item{xlim}{range of x values to graph}

\item{...}{Additional arguments passed to \code{\link[ggplot2:geom_function]{ggplot2::geom_function()}}.
\code{fun}, \code{n}, and \code{args} are set internally and cannot be passed.}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Graph an antibody decay curve model
}
\details{
\subsection{Fixed \code{...} arguments}{

The arguments \code{fun}, \code{n}, and \code{args} are set internally and cannot be
overridden via \code{...}. Passing them will trigger an informative error.
}

\subsection{\code{n_curves} and \code{iters_to_graph}}{

In most cases, \code{object} will contain too many rows of MCMC
samples for all of these samples to be plotted at once.
\itemize{
\item Setting the  \code{n_curves} argument to a value smaller than the
number of rows in \code{curve_params} will cause this function to select
the first \code{n_curves} rows to graph.
\item Setting \code{n_curves} larger than the number of rows in ` will
result all curves being plotted.
\item If the user directly specifies the \code{iters_to_graph} argument,
then \code{n_curves} has no effect.
}
}
}
\examples{
\donttest{
library(dplyr) # loads the `\%>\%` operator and `dplyr::filter()`

curve <-
  typhoid_curves_nostrat_100 \%>\%
  filter(antigen_iso == ("HlyE_IgG")) \%>\%
  serocalculator:::plot_curve_params_one_ab()

  curve
}
}
\keyword{internal}
