% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_decay_curve.R
\name{plot_decay_curve}
\alias{plot_decay_curve}
\title{Plot a decay function}
\usage{
plot_decay_curve(decay_function, ..., xmax = 100, ymax = NA, title = "")
}
\arguments{
\item{decay_function}{a function with first argument \code{t}}

\item{...}{parameters passed to \code{decay_function}}

\item{xmax}{upper limit of x axis}

\item{ymax}{upper limit of y axis}

\item{title}{plot title}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Plot a decay function
}
\examples{
plot_decay_curve(antibody_decay_curve)
}
