% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.icc_seroincidence.by.R
\name{print.icc_seroincidence.by}
\alias{print.icc_seroincidence.by}
\title{Print method for stratified ICC results}
\usage{
\method{print}{icc_seroincidence.by}(x, ...)
}
\arguments{
\item{x}{an object of class \code{icc_seroincidence.by}}

\item{...}{unused}
}
\value{
invisible x
}
\description{
Print method for stratified ICC results
}
