% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_robust_se.R
\name{print.icc_seroincidence}
\alias{print.icc_seroincidence}
\title{Print method for ICC results}
\usage{
\method{print}{icc_seroincidence}(x, ...)
}
\arguments{
\item{x}{an object of class \code{icc_seroincidence} or \code{icc_seroincidence.by}}

\item{...}{unused}
}
\value{
invisible x
}
\description{
Print method for ICC results
}
