% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.posterior_predictive_check}
\alias{print.posterior_predictive_check}
\title{Print Posterior Predictive Check Results}
\usage{
\method{print}{posterior_predictive_check}(x, ...)
}
\arguments{
\item{x}{a \code{posterior_predictive_check} object}

\item{...}{unused}
}
\description{
Print Posterior Predictive Check Results
}
