% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.seroincidence_residuals}
\alias{print.seroincidence_residuals}
\title{Print Seroincidence Residuals}
\usage{
\method{print}{seroincidence_residuals}(x, ...)
}
\arguments{
\item{x}{a \code{seroincidence_residuals} tibble}

\item{...}{passed to \code{\link[tibble:print.tbl_df]{tibble::print.tbl_df()}}}
}
\description{
Print Seroincidence Residuals
}
