% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.seroincidence.R
\name{print.seroincidenceList}
\alias{print.seroincidenceList}
\title{Print Method for Seroincidence Object}
\usage{
\method{print}{seroincidenceList}(x, ...)
}
\arguments{
\item{x}{A list containing output of function \code{\link[=estimateSeroincidence]{estimateSeroincidence()}}.}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
Custom \code{\link[=print]{print()}} function to show output of the seroincidence calculator \code{\link[=estimateSeroincidence]{estimateSeroincidence()}}.
}
\examples{

\dontrun{
# estimate seroincidence
seroincidence <- estimateSeroincidence(...)

# print the seroincidence object to the console
print(seroincidence)

# or simply type (appropriate print method will be invoked automatically)
seroincidence
}

}
