% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.seroincidence.R
\name{print.summary.seroincidence}
\alias{print.summary.seroincidence}
\title{Print Method for Seroincidence Summary Object}
\usage{
\method{print}{summary.seroincidence}(x, ...)
}
\arguments{
\item{x}{A list containing output of function \code{\link{summary.seroincidence}}.}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
Custom \code{\link{print}} function to show output of the seroincidence summary \code{\link{summary.seroincidence}}.
}
\examples{

\dontrun{

# estimate seroincidence
seroincidence <- estimateSeroincidence(...)

# calculate summary statistics for the seroincidence object
seroincidenceSummary <- summary(seroincidence)

# print the summary of seroincidence object to the console
print(seroincidenceSummary)

# or simply type (appropriate print method will be invoked automatically)
seroincidenceSummary
}

}

