% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.cs-utils.r
\name{row_longitudinal_parameter}
\alias{row_longitudinal_parameter}
\title{extract a row from longitudinal parameter set}
\usage{
row_longitudinal_parameter(age, antigen_isos, nmc, npar, ...)
}
\arguments{
\item{age}{age at infection}

\item{antigen_isos}{antigen isotypes}

\item{nmc}{mcmc sample to use}

\item{npar}{number of parameters}

\item{...}{passed to \code{simpar()}}
}
\value{
an array of parameters:
c(y0,b0,mu0,mu1,c1,alpha,shape)
}
\description{
take a random sample from longitudinal parameter set
given age at infection, for a  list of antibodies
}
