% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sees_baseline_data.R
\docType{data}
\name{sees_crossSectional_baseline_allCountries}
\alias{sees_crossSectional_baseline_allCountries}
\title{Cross-sectional Typhoid data}
\usage{
sees_crossSectional_baseline_allCountries
}
\description{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
