% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sees_typhoid_ests_strat.R
\docType{data}
\name{sees_typhoid_ests_strat}
\alias{sees_typhoid_ests_strat}
\title{Example \code{"seroincidence.by"} object}
\format{
An object of class \code{seroincidence.by} (inherits from \code{list}) of length 9.
}
\source{
\code{serocalculator/data-raw/sees_typhoid_ests_strat.R}
}
\usage{
sees_typhoid_ests_strat
}
\description{
Typhoid seroconversion rate estimates by country and age category
from the SEES study.
}
\keyword{datasets}
