% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serocalculator-package.R
\docType{package}
\name{serocalculator-package}
\alias{serocalculator}
\alias{serocalculator-package}
\title{serocalculator: Estimating Infection Rates from Serological Data}
\description{
Translates antibody levels measured in cross-sectional population samples into estimates of the frequency with which seroconversions (infections) occur in the sampled populations. Replaces the previous `seroincidence` package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ucd-serg.github.io/serocalculator/}
  \item \url{https://github.com/UCD-SERG/serocalculator}
  \item Report bugs at \url{https://github.com/UCD-SERG/serocalculator/issues}
}

}
\author{
\strong{Maintainer}: Kristina Lai \email{kwlai@ucdavis.edu}

Authors:
\itemize{
  \item Peter Teunis \email{p.teunis@emory.edu} (Author of the method and original code.) [copyright holder]
  \item Chris Orwa
  \item Kristen Aiemjoy \email{kaiemjoy@ucdavis.edu}
  \item Douglas Ezra Morrison \email{demorrison@ucdavis.edu}
}

Other contributors:
\itemize{
  \item Kwan Ho Lee \email{ksjlee@ucdavis.edu} [contributor]
}

}
\keyword{internal}
