% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serocalculator_example.R
\name{serocalculator_example}
\alias{serocalculator_example}
\title{Get path to an example file}
\usage{
serocalculator_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\value{
a \link{character} string providing
the path to the file specified by \code{file},
or a vector or available files if \code{file = NULL}.
}
\description{
The \link{serocalculator} package comes bundled with a number of sample files
in its \code{inst/extdata} directory.
This \code{serocalculator_example()} function make those sample files
easy to access.
}
\details{
Adapted from \code{\link[readr:readr_example]{readr::readr_example()}} following the guidance in
\url{https://r-pkgs.org/data.html#sec-data-example-path-helper}.
}
\examples{
serocalculator_example()
serocalculator_example("example_pop_data.csv")
}
