% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop_data_multi.R
\name{sim.cs.multi}
\alias{sim.cs.multi}
\title{Simulate multiple data sets}
\usage{
sim.cs.multi(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{sim.cs.multi()} was renamed to \code{\link[=sim_pop_data_multi]{sim_pop_data_multi()}}
to create a more consistent API.
}
\keyword{internal}
