% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pop_data.R
\name{sim.cs}
\alias{sim.cs}
\title{Simulate a cross-sectional serosurvey with noise}
\usage{
sim.cs(n.smpl, age.rng, n.mc, renew.params, add.noise, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{sim.cs()} was renamed to \code{\link[=sim_pop_data]{sim_pop_data()}} to create a more
consistent API.
}
\keyword{internal}
