% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_seroincidence.R
\name{simulate_seroincidence}
\alias{simulate_seroincidence}
\title{Simulate seroincidence}
\usage{
simulate_seroincidence(
  nrep = 100,
  n_sim = 10,
  sr_params,
  renew.params = TRUE,
  antibodies = c("HlyE_IgA", "HlyE_IgG"),
  lambda = 0.01,
  lifespan = c(0, 10),
  dlims = rbind(HlyE_IgA = c(min = 0, max = 0.5), HlyE_IgG = c(min = 0, max = 0.5)),
  cond = tibble(antigen_iso = c("HlyE_IgG", "HlyE_IgA"), nu = c(0.5, 0.5), eps = c(0, 0),
    y.low = c(1, 1), y.high = c(5e+06, 5e+06))
)
}
\arguments{
\item{nrep}{number of samples per simulated dataset}

\item{n_sim}{number of simulations to run}

\item{sr_params}{Curve parameters}

\item{renew.params}{whether to sample new parameters for each infection
in a simulated individual's longitudinal data}

\item{antibodies}{which antibodies to simulate}

\item{lambda}{Simulated incidence rate per person-year}

\item{lifespan}{range of ages to simulate}

\item{dlims}{biologic noise distribution}

\item{cond}{Noise parameters}
}
\value{
a \link{list} of simulation results
}
\description{
Simulate seroincidence
}
\examples{
simulate_seroincidence(sr_params = typhoid_curves_nostrat_100)
}
\keyword{internal}
