% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{strata.default}
\alias{strata.default}
\title{Extract the \code{Strata} attribute from an object, if present}
\usage{
\method{strata}{default}(x)
}
\arguments{
\item{x}{any R object}
}
\value{
\itemize{
\item a \code{\link[tibble:tibble]{tibble::tibble()}} with strata in rows, or
\item \code{NULL} if \code{x} does not have a \code{"strata"} attribute
}
}
\description{
Extract the \code{Strata} attribute from an object, if present
}
\keyword{internal}
