% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{strata}
\alias{strata}
\title{Extract \code{Strata} metadata from an object}
\usage{
strata(x)
}
\arguments{
\item{x}{an object}
}
\value{
the strata metadata of \code{x}
}
\description{
Generic method for extracting strata metadata from objects.
See \code{\link[=strata.default]{strata.default()}}
}
