% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.seroincidence.ests.R
\name{[.seroincidence.by}
\alias{[.seroincidence.by}
\title{Extract or replace parts of a \code{seroincidence.by} object}
\usage{
\method{[}{seroincidence.by}(x, i, ...)
}
\arguments{
\item{x}{the object to subset/replace elements of}

\item{i}{the indices to subset/replace}

\item{...}{passed to \verb{[.list}}
}
\value{
the subset specified
}
\description{
Extract or replace parts of a \code{seroincidence.by} object
}
