% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.seroincidence.ests.R
\name{[.seroincidence.ests}
\alias{[.seroincidence.ests}
\title{Extract or replace parts of a \code{seroincidence.ests} object}
\usage{
\method{[}{seroincidence.ests}(x, i, ...)
}
\value{
the subset specified
}
\description{
Extract or replace parts of a \code{seroincidence.ests} object
}
