% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_fit.R
\name{summary.seroincidence.est}
\alias{summary.seroincidence.est}
\title{summarize a fitted incidence model}
\usage{
\method{summary}{seroincidence.est}(object, coverage = 0.95, ...)
}
\arguments{
\item{object}{a \code{\link[=list]{list()}}, outputted by \code{\link[stats:nlm]{stats::nlm()}} or \code{\link[=find.MLE]{find.MLE()}}}

\item{coverage}{desired confidence interval coverage probability}

\item{...}{unused}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}}; see \code{\link[stats:nlm]{stats::nlm()}} for details on \code{code} variable
}
\description{
summarize a fitted incidence model
}
