% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.seroincidence.R
\name{summary.seroincidence.ests}
\alias{summary.seroincidence.ests}
\title{Summary Method for \code{"seroincidence.ests"} Objects}
\usage{
\method{summary}{seroincidence.ests}(
  object,
  confidence_level = 0.95,
  showDeviance = TRUE,
  showConvergence = TRUE,
  ...
)
}
\arguments{
\item{object}{A dataframe containing output of function \code{\link[=est.incidence.by]{est.incidence.by()}}.}

\item{confidence_level}{desired confidence interval coverage probability}

\item{showDeviance}{Logical flag (\code{FALSE}/\code{TRUE}) for reporting deviance
(-2*log(likelihood) at estimated seroincidence. Default = \code{TRUE}.}

\item{showConvergence}{Logical flag (\code{FALSE}/\code{TRUE}) for reporting convergence (see
help for \code{\link[=optim]{optim()}} for details). Default = \code{FALSE}.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
A \code{summary.seroincidence.ests} object, which is a \link[dplyr:reexports]{dplyr::tibble}, with the following columns:
\itemize{
\item \code{incidence.rate} maximum likelihood estimate of \code{lambda} (seroincidence)
\item \code{CI.lwr} lower confidence bound for lambda
\item \code{CI.upr} upper confidence bound for lambda
\item \code{Deviance} (included if \code{showDeviance = TRUE}) Negative log likelihood (NLL) at estimated (maximum likelihood)
\code{lambda})
\itemize{
\item \code{nlm.exit.code} (included if \code{showConvergence = TRUE}) Convergence information returned by \code{\link[stats:nlm]{stats::nlm()}}
The object also has the following metadata (accessible through \code{\link[base:attr]{base::attr()}}):
}
\item \code{antigen_isos} Character vector with names of input antibodies used in \code{\link[=est.incidence.by]{est.incidence.by()}}
\item \code{Strata} Character with names of strata used in \code{\link[=est.incidence.by]{est.incidence.by()}}
}
}
\description{
Calculate seroincidence from output of the seroincidence calculator
\code{\link[=est.incidence.by]{est.incidence.by()}}.
}
\examples{

\dontrun{
# estimate seroincidence
seroincidence <- est.incidence.by(...)

# calculate summary statistics for the seroincidence object
seroincidenceSummary <- summary(seroincidence)
}

}
