% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typhoid_controls.R
\docType{data}
\name{typhoid_controls}
\alias{typhoid_controls}
\title{Example cross-sectional data for typhoid (\code{HlyE_IgA} and \code{HlyE_IgG})}
\format{
The measurements are in the variable \code{elisa}. Variable \code{nu} provides an estimate of a conditional noise parameter.
}
\usage{
typhoid_controls
}
\description{
Data frame of ELISA assay results, by \code{pop} (\code{"CA Facts"}, \code{"MGH"}), \code{antigen_iso} (\code{"HlyE_IgA"}, \code{"HlyE_IgG"}), and \code{Age} (in years).
}
\examples{

# Print the data:
typhoid_controls

# Plot the data
library(ggplot2)
typhoid_controls |>
  ggplot(aes(x = Age, y = elisa, col = antigen_iso)) +
  geom_point() +
  geom_smooth()

}
