% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typhoid_noise_params.R
\docType{data}
\name{typhoid_noise_params}
\alias{typhoid_noise_params}
\title{Example noise parameters for HlyE_IgA and HlyE_IgG, by country}
\format{
The parameters are:
\itemize{
\item \code{llod}
\item \code{nu}
\item \code{y.high}
\item \code{eps}
}
}
\usage{
typhoid_noise_params
}
\description{
Data frame of example noise parameters, by \code{Country} (\code{"MGH"}, \code{"bangladesh"}, \code{"nepal"})
and \code{antigen_iso} (\code{"HlyE_IgA"}, \code{"HlyE_IgG"}).
}
\examples{

# Show the data:
typhoid_noise_params

}
