% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab.R
\name{ab}
\alias{ab}
\title{kinetics of the antibody (ab) response (power function decay)}
\usage{
ab(t, par, ...)
}
\arguments{
\item{t}{age at infection?}

\item{par}{parameters}

\item{...}{arguments passed to \code{baseline()}}
}
\value{
a \code{\link[=matrix]{matrix()}}
}
\description{
kinetics of the antibody (ab) response (power function decay)
}
