% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sims.R
\name{analyze_sims}
\alias{analyze_sims}
\title{Analyze simulation results}
\usage{
analyze_sims(data, true_lambda = attr(data, "true_lambda"))
}
\arguments{
\item{data}{a tibble with}

\item{true_lambda}{data-generating incidence rate}
}
\value{
a \link{list}
}
\description{
Analyze simulation results
}
