% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sims.R
\name{analyze_sims}
\alias{analyze_sims}
\title{Analyze simulation results}
\usage{
analyze_sims(data)
}
\arguments{
\item{data}{a \link[tibble:tibble]{tibble::tibble} with columns:
\itemize{
\item \code{lambda.sim},
\item \code{incidence.rate},
\item \code{SE},
\item \code{CI.lwr},
\item \code{CI.upr}
for example, as produced by \code{\link[=summary.seroincidence.by]{summary.seroincidence.by()}} with
\code{lambda.sim} as a stratifying variable
}}
}
\value{
a \link{list}
}
\description{
Analyze simulation results
}
