% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antibody_decay_function.R
\name{antibody_decay_curve}
\alias{antibody_decay_curve}
\title{Graph antibody decay function with natural parameters}
\usage{
antibody_decay_curve(
  t,
  y0 = 0.74916052,
  b0 = 1,
  mu_b = 0.18432798,
  mu_y = 0.36853621,
  gamma = 0.0013040664,
  alpha = 2.192627e-05,
  rho = 2.9621994
)
}
\arguments{
\item{t}{time since infection}

\item{y0}{initial antibody concnetration}

\item{b0}{initial bacteria concentration}

\item{mu_b}{pathogen reproduction rate}

\item{mu_y}{antibody reproduction rate}

\item{gamma}{bacteria destruction rate per antibody}

\item{alpha}{antibody decay rate}

\item{rho}{antibody decay rate power function parameter}
}
\value{
a \code{\link[=numeric]{numeric()}} vector
}
\description{
Graph antibody decay function with natural parameters
}
\examples{
library(ggplot2)
ggplot() + geom_function(fun = antibody_decay_curve) + xlim(1,100)
}
