% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.curve_params.R
\name{autoplot.curve_params}
\alias{autoplot.curve_params}
\title{graph antibody decay curves by antigen isotype}
\usage{
\method{autoplot}{curve_params}(object, antigen_isos = unique(object$antigen_iso), ...)
}
\arguments{
\item{object}{a \code{\link[=data.frame]{data.frame()}} of curve parameters (one or more MCMC samples)}

\item{antigen_isos}{antigen isotypes to analyze
(can subset \code{curve_params})}

\item{...}{
  Arguments passed on to \code{\link[=graph_seroresponse_model_1]{graph_seroresponse_model_1}}
  \describe{
    \item{\code{ncol}}{how many columns of subfigures to use in panel plot}
  }}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
graph antibody decay curves by antigen isotype
}
\examples{
\donttest{
library(dplyr)
library(ggplot2)
library(magrittr)

curve <-
  serocalculator_example("example_curve_params.csv") |>
  read.csv() |>
  as_curve_params() |>
  filter(antigen_iso \%in\% c("HlyE_IgA", "HlyE_IgG")) |>
  autoplot()

curve
}
}
