% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.curve_params.R
\name{autoplot.curve_params}
\alias{autoplot.curve_params}
\title{Graph antibody decay curves by antigen isotype}
\usage{
\method{autoplot}{curve_params}(object, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=graph.curve.params]{graph.curve.params}}
  \describe{
    \item{\code{curve_params}}{a \code{\link[=data.frame]{data.frame()}} containing MCMC samples of antibody decay curve parameters}
    \item{\code{verbose}}{verbose output}
    \item{\code{show_all_curves}}{whether to show individual curves under quantiles}
    \item{\code{alpha_samples}}{\code{alpha} parameter passed to \link[ggplot2:geom_path]{ggplot2::geom_line}
(has no effect if \code{show_all_curves = FALSE})}
    \item{\code{show_quantiles}}{whether to show point-wise (over time) quantiles}
  }}

\item{antigen_isos}{antigen isotypes to analyze
(can subset \code{curve_params})}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Graph antibody decay curves by antigen isotype
}
\details{
Currently, the backend for this method is \code{\link[=graph.curve.params]{graph.curve.params()}}.
Previously, the backend for this method was \code{\link[=graph_seroresponse_model_1]{graph_seroresponse_model_1()}}.
That function is still available if preferred.
}
\examples{
\donttest{
library(dplyr)
library(ggplot2)
library(magrittr)

curve <-
  serocalculator_example("example_curve_params.csv") |>
  read.csv() |>
  as_sr_params() |>
  filter(antigen_iso \%in\% c("HlyE_IgA", "HlyE_IgG")) |>
  autoplot()

curve
}
}
