% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.pop_data.R
\name{autoplot.pop_data}
\alias{autoplot.pop_data}
\title{Plot distribution of antibodies}
\usage{
\method{autoplot}{pop_data}(pop_data, log = FALSE, strata = NULL)
}
\arguments{
\item{pop_data}{A \code{pop_data} object (from \code{\link[=load_pop_data]{load_pop_data()}})}

\item{log}{whether to show antibody responses on logarithmic scale}

\item{strata}{the name of a variable in \code{pop_data} to stratify by (or \code{NULL} for no stratification)}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot} object
}
\description{
Plot distribution of antibodies
}
