% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.seroincidence.ests.R
\name{autoplot.seroincidence.by}
\alias{autoplot.seroincidence.by}
\title{Plot \code{seroincidence.by} log-likelihoods}
\usage{
\method{autoplot}{seroincidence.by}(object, ncol = min(3, length(object)), ...)
}
\arguments{
\item{object}{a '"seroincidence.by"' object (from \code{\link[=est.incidence.by]{est.incidence.by()}})}

\item{ncol}{number of columns to use for panel of plots}

\item{...}{
  Arguments passed on to \code{\link[=autoplot.seroincidence]{autoplot.seroincidence}}
  \describe{
    \item{\code{log_x}}{should the x-axis be on a logarithmic scale (\code{TRUE}) or linear scale (\code{FALSE}, default)?}
  }}
}
\value{
an object of class \code{"ggarrange"}, which is a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} or a \code{\link[=list]{list()}} of \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}s.
}
\description{
Plots log-likelihood curves by stratum, for \code{seroincidence.by} objects
}
