% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.seroincidence.est.R
\name{autoplot.seroincidence}
\alias{autoplot.seroincidence}
\title{Plot the log-likelihood curve for the incidence rate estimate}
\usage{
\method{autoplot}{seroincidence}(object, log_x = FALSE, ...)
}
\arguments{
\item{object}{a \code{seroincidence} object (from \code{\link[=est.incidence]{est.incidence()}})}

\item{log_x}{should the x-axis be on a logarithmic scale (\code{TRUE}) or linear scale (\code{FALSE}, default)?}

\item{...}{unused}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
\description{
Plot the log-likelihood curve for the incidence rate estimate
}
\examples{
 #generate cross-sectional data
csdata <- sim.cs(
  curve_params = dmcmc,
  lambda = lambda,
  n.smpl = nrep,
  age.rng = lifespan,
  antigen_isos = antibodies,
 n.mc = 0,
 renew.params = TRUE,
 add.noise = TRUE,
 noise_limits = dlims,
 format = "long"
)

#load in longitudinal parameters
dmcmc =
  "https://osf.io/download/rtw5k" |>
  load_curve_params()

#Load noise params
  cond <- tibble(
  antigen_iso = c("HlyE_IgG", "HlyE_IgA"),
  nu = c(0.5, 0.5),                          # Biologic noise (nu)
  eps = c(0, 0),                             # M noise (eps)
  y.low = c(1, 1),                           # low cutoff (llod)
  y.high = c(5e6, 5e6))                      # high cutoff (y.high)

#Calculate seroincidence
est1 = est.incidence(
  pop_data = csdata,
  curve_params = dmcmc,
  noise_params = cond,
  lambda_start = .1,
  build_graph = T,
  verbose = T, # print updates as the function runs
  print_graph = F, # display the log-likelihood curve while `est.incidence()` is running
  antigen_isos = antibodies)

#plot the log-likelihood curve
  autoplot(est1)

}
