% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.summary.seroincidence.by.R
\name{autoplot.summary.seroincidence.by}
\alias{autoplot.summary.seroincidence.by}
\title{Plot method for \code{summary.seroincidence.by} objects}
\usage{
\method{autoplot}{summary.seroincidence.by}(object, type, ...)
}
\arguments{
\item{object}{a \code{summary.seroincidence.by} object
(generated by applying the \code{summary()}
method to the output of \code{\link[=estimate_scr_by]{estimate_scr_by()}}).}

\item{type}{\link{character} string indicating which type of plot to generate.
The implemented options are \code{"scatter"} and \code{"bar"}.}

\item{...}{
  Arguments passed on to \code{\link[=strat_ests_scatterplot]{strat_ests_scatterplot}}
  \describe{
    \item{\code{xvar}}{the name of a stratifying variable in \code{object}}
    \item{\code{alpha}}{transparency for the points in the graph
(1 = no transparency, 0 = fully transparent)}
    \item{\code{shape}}{shape argument for \code{geom_point()}}
    \item{\code{dodge_width}}{width for jitter}
    \item{\code{CIs}}{\link{logical}, if \code{TRUE}, add CI error bars}
    \item{\code{color_var}}{\link{character} which variable in \code{object} to use
to determine point color}
    \item{\code{group_var}}{\link{character} which variable in \code{object} to use
to connect points with lines (\code{NULL} for no lines)}
  }}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot method for \code{summary.seroincidence.by} objects
}
\examples{

library(dplyr)
library(ggplot2)

xs_data <-
  sees_pop_data_pk_100

curve <-
  typhoid_curves_nostrat_100 \%>\%
  filter(antigen_iso \%in\% c("HlyE_IgA", "HlyE_IgG"))

noise <-
  example_noise_params_pk

est2 <- estimate_scr_by(
  strata = c("catchment", "ageCat"),
  pop_data = xs_data,
  curve_params = curve,
  noise_params = noise,
  curve_strata_varnames= NULL,
  noise_strata_varnames = NULL,
  antigen_isos = c("HlyE_IgG", "HlyE_IgA"),
  num_cores = 2 # Allow for parallel processing to decrease run time
)

est2sum <- summary(est2)

est2sum |> strat_ests_scatterplot(
    type ="scatter",
    xvar = "ageCat",
    color_var = "catchment",
    CIs = TRUE,
    group_var = "catchment")

}
