% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.summary.seroincidence.by.R
\name{autoplot.summary.seroincidence.by}
\alias{autoplot.summary.seroincidence.by}
\title{Plot method for \code{summary.seroincidence.by} objects}
\usage{
\method{autoplot}{summary.seroincidence.by}(
  object,
  type = "scatter",
  xvar,
  fill_var,
  alpha = 0.7,
  shape = 1,
  dodge_width = 0.001,
  CIs = FALSE,
  title = "Seroconversion Rate by Group",
  xlab = "Seroconversion rate per 1000 person-years",
  ylab = "Stratification Variable",
  fill_lab = fill_var,
  fill_palette = NULL,
  ...
)
}
\arguments{
\item{object}{a \code{summary.seroincidence.by} object
(generated by applying the \code{summary()}
method to the output of \code{\link[=estimate_scr_by]{estimate_scr_by()}}).}

\item{type}{an option to choose type of chart:
the current options are \code{"scatter"} or \code{"bar"}}

\item{xvar}{the name of a stratifying variable in \code{object}}

\item{fill_var}{the name of the variable used for fill color (e.g., "Country")}

\item{alpha}{transparency for the points in the graph
(1 = no transparency, 0 = fully transparent)}

\item{shape}{shape argument for \code{geom_point()}}

\item{dodge_width}{width for jitter}

\item{CIs}{\link{logical}, if \code{TRUE}, add CI error bars}

\item{title}{a title for the final plot}

\item{xlab}{a label for the x axis of the final plot}

\item{ylab}{the name of a second stratifying variable in \code{object} (e.g., "catchment")}

\item{fill_lab}{fill label}

\item{fill_palette}{define the color palette for bar fill}

\item{...}{unused}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot method for \code{summary.seroincidence.by} objects
}
\examples{

library(dplyr)
library(ggplot2)

xs_data <-
  sees_pop_data_pk_100

curve <-
  typhoid_curves_nostrat_100 \%>\%
  filter(antigen_iso \%in\% c("HlyE_IgA", "HlyE_IgG"))

noise <-
  example_noise_params_pk

est2 <- estimate_scr_by(
  strata = c("catchment", "ageCat"),
  pop_data = xs_data,
  curve_params = curve,
  noise_params = noise,
  curve_strata_varnames= NULL,
  noise_strata_varnames = NULL,
  antigen_isos = c("HlyE_IgG", "HlyE_IgA"),
  num_cores = 2 # Allow for parallel processing to decrease run time
)

est2sum <- summary(est2)

autoplot(est2sum, "ageCat", type = "scatter", fill_var = "catchment")
autoplot(est2sum, "ageCat", type = "bar", fill_var = "catchment")

}
