% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pop_data.R
\name{check_pop_data}
\alias{check_pop_data}
\title{Check the formatting of a cross-sectional antibody survey dataset.}
\usage{
check_pop_data(pop_data)
}
\arguments{
\item{pop_data}{dataset to check}
}
\value{
NULL (invisibly)
}
\description{
Check the formatting of a cross-sectional antibody survey dataset.
}
\examples{
library(dplyr)

# Import cross-sectional data from OSF and rename required variables
xs_data <- load_pop_data("https://osf.io/download//n6cp3/") \%>\%
  clean_pop_data() \%>\%
  check_pop_data()

}
