% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pop_data.R
\name{check_pop_data}
\alias{check_pop_data}
\title{Check the formatting of a cross-sectional antibody survey dataset.}
\usage{
check_pop_data(pop_data)
}
\arguments{
\item{pop_data}{dataset to check}
}
\value{
NULL (invisibly)
}
\description{
Check the formatting of a cross-sectional antibody survey dataset.
}
