% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pop_data.R
\name{check_pop_data}
\alias{check_pop_data}
\title{Check the formatting of a cross-sectional antibody survey dataset.}
\usage{
check_pop_data(pop_data, antigen_isos)
}
\arguments{
\item{antigen_isos}{}

\item{data}{a \link{data.frame}}
}
\value{
NULL (invisibly)
}
\description{
Check the formatting of a cross-sectional antibody survey dataset.
}
