% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_strata.R
\name{check_strata}
\alias{check_strata}
\title{Check a \code{pop_data} object for requested strata variables}
\usage{
check_strata(pop_data, strata)
}
\arguments{
\item{pop_data}{a \code{pop_data} object}

\item{strata}{a \link{character} vector}
}
\value{
\link{NULL}, invisibly
}
\description{
Check a \code{pop_data} object for requested strata variables
}
\examples{
sees_pop_data_pk_100 |>
  check_strata(strata = c("ag", "catch", "Count"))
}
\keyword{internal}
