% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_array.R
\name{df.to.array}
\alias{df.to.array}
\title{Convert a data.frame (or tibble) into a multidimensional array}
\usage{
df.to.array(df, dim_var_names, value_var_name = "value")
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{df.to.array()} was renamed to \code{df_to_array()} to create a more
consistent API.
}
\keyword{internal}
