% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{.nll}
\alias{.nll}
\title{Calculate log-likelihood}
\usage{
.nll(
  stratumData,
  antibodies,
  params,
  censorLimits,
  ivc = FALSE,
  m = 0,
  par0,
  start
)
}
\arguments{
\item{ivc}{If `ivc = TRUE`, the biomarker data are interval-censored.}

\item{start}{}
}
\value{

}
\description{
Calculate log-likelihood
}
