% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{.nll}
\alias{.nll}
\title{Calculate negative log-likelihood}
\usage{
.nll(log.lambda, ...)
}
\arguments{
\item{log.lambda}{natural logarithm of incidence rate}

\item{...}{
  Arguments passed on to \code{\link[=llik]{llik}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
the negative log-likelihood of the data with the current parameter values
}
\description{
Same as \code{\link[=llik]{llik()}}, except negated and requiring lambda on log scale (used in combination with \code{nlm()}, to ensure that the optimization search doesn't stray into negative values of \code{lambda}).
}
\keyword{internal}
