% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{.nll}
\alias{.nll}
\title{Calculate negative log-likelihood}
\usage{
.nll(log.lambda, ...)
}
\arguments{
\item{log.lambda}{log.lambda}

\item{...}{
  Arguments passed on to \code{\link[=llik]{llik}}
  \describe{
    \item{\code{data}}{Data frame with cross-sectional serology data per antibody and age, and additional columns}
    \item{\code{antigen_isos}}{Character vector with one or more antibody names. Values must match \code{data}.}
    \item{\code{curve_params}}{List of data frames of all longitudinal parameters. Each data frame contains
Monte Carlo samples for each antibody type.}
    \item{\code{noise_params}}{a \code{\link[=list]{list()}} (or \code{\link[=data.frame]{data.frame()}}, or \code{\link[=tibble]{tibble()}}) containing noise parameters}
    \item{\code{verbose}}{logical: if TRUE, print verbose log information to console}
  }}
}
\value{
the negative log-likelihood of the data with the current parameter values
}
\description{
Calculate negative log-likelihood
}
\details{
Same as \code{\link[=.nll]{.nll()}}, except negated
}
