% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{.nll}
\alias{.nll}
\title{Calculate log-likelihood}
\usage{
.nll(
  stratumData,
  antibodies,
  params,
  censorLimits,
  ivc = FALSE,
  m = 0,
  par0,
  start
)
}
\arguments{
\item{stratumData}{Data frame with cross-sectional serology data per antibody and age, and additional
columns, for one stratum}

\item{antibodies}{Character vector with one or more antibody names. Values must match \code{data}.}

\item{params}{List of data frames of all longitudinal parameters. Each data frame contains
Monte Carlo samples for each antibody type.}

\item{censorLimits}{List of cutoffs for one or more named antibody types (corresponding to
\code{stratumData}).}

\item{ivc}{If `ivc = TRUE`, the biomarker data are interval-censored.}

\item{m}{[not sure what this is]}

\item{par0}{List of parameters for the (lognormal) distribution of antibody concentrations
for true seronegatives (i.e. those who never seroconverted), by named antibody type
(corresponding to \code{data}).}

\item{start}{starting value for \code{log(lambda)}. Value of -6 corresponds roughly to 1 day
(log(1/365.25)), -4 corresponds roughly to 1 week (log(7 / 365.25)). Default = -6.}
}
\value{
the log-likelihood of the data with the current parameter values
}
\description{
Calculate log-likelihood
}
