% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{.nll}
\alias{.nll}
\title{Calculate negative log-likelihood}
\usage{
.nll(log.lambda, ...)
}
\arguments{
\item{log.lambda}{natural logarithm of incidence rate}

\item{...}{
  Arguments passed on to \code{\link[=llik]{llik}}
  \describe{
    \item{\code{data}}{Data frame with cross-sectional serology data per antibody and age, and additional columns}
    \item{\code{antigen_isos}}{Character vector listing one or more antigen isotypes. Values must match \code{data}.}
    \item{\code{curve_params}}{List of data frames of all longitudinal parameters. Each data frame contains Monte Carlo samples for each antibody type.}
    \item{\code{noise_params}}{a \code{\link[=data.frame]{data.frame()}} (or \code{\link[tibble:tibble]{tibble::tibble()}}) containing the following variables, specifying noise parameters for each antigen isotype:
\itemize{
\item \code{antigen_iso}: antigen isotype whose noise parameters are being specified on each row
\item \code{nu}: biological noise
\item \code{eps}: measurement noise
\item \code{y.low}: lower limit of detection for the current antigen isotype
\item \code{y.high}: upper limit of detection for the current antigen isotype
}}
    \item{\code{verbose}}{logical: if TRUE, print verbose log information to console}
  }}
}
\value{
the negative log-likelihood of the data with the current parameter values
}
\description{
Calculate negative log-likelihood
}
\details{
Same as \code{\link[=llik]{llik()}}, except negated and requiring lambda on log scale (used in combination with \code{nlm()}, to ensure that the optimization search doesn't stray into negative values of \code{lambda}).
}
