% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{.nll}
\alias{.nll}
\title{Calculate log-likelihood}
\usage{
.nll(`log(lambda)`, data, antibodies, lnparams, noise_params)
}
\arguments{
\item{data}{Data frame with cross-sectional serology data per antibody and age, and additional columns, for one stratum}

\item{antibodies}{Character vector with one or more antibody names. Values must match \code{data}.}

\item{lnparams}{List of data frames of all longitudinal parameters. Each data frame contains
Monte Carlo samples for each antibody type.}

\item{noise_params}{a \code{\link[=list]{list()}} (or \code{\link[=data.frame]{data.frame()}}, or \code{\link[=tibble]{tibble()}}) containing noise parameters}

\item{`log(lambda)`}{natural logarithm of incidence parameter, in log(years). Value of -6 corresponds roughly to 1 day (log(1/365.25)), -4 corresponds roughly to 1 week (log(7 / 365.25)). Default = -6.}
}
\value{
the log-likelihood of the data with the current parameter values
}
\description{
Calculate log-likelihood
}
