% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.incidence.by.R
\name{est.incidence.by}
\alias{est.incidence.by}
\title{Estimate Seroincidence}
\usage{
est.incidence.by(
  data,
  curve_params,
  noise_params,
  strata = "",
  curve_strata_varnames = strata,
  noise_strata_varnames = strata,
  numCores = 1L,
  antigen_isos = unique(pull(data, "antigen_iso")),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame with cross-sectional serology data per antibody and age, and additional columns to identify possible \code{strata}.}

\item{curve_params}{List of data frames of all longitudinal parameters. Each data frame contains
Monte Carlo samples for each antibody type.}

\item{noise_params}{a \code{\link[=list]{list()}} (or \code{\link[=data.frame]{data.frame()}}, or \code{\link[=tibble]{tibble()}}) containing noise parameters}

\item{strata}{Character vector of stratum-defining variables. Values must be variable names in \code{data}. Default = "".}

\item{curve_strata_varnames}{A subset of \code{strata}. Values must be variable names in \code{curve_params}. Default = "".}

\item{noise_strata_varnames}{A subset of \code{strata}. Values must be variable names in \code{noise_params}. Default = "".}

\item{numCores}{Number of processor cores to use for calculations when computing by strata. If set to more than 1 and package \pkg{parallel} is available, then the computations are executed in parallel. Default = 1L.}

\item{antigen_isos}{Character vector with one or more antibody names. Values must match \code{data}}

\item{verbose}{logical: if TRUE, print verbose log information to console}

\item{...}{
  Arguments passed on to \code{\link[=.optNll]{.optNll}}
  \describe{
    \item{\code{lambda.start}}{starting guess for incidence rate, in years/event.}
    \item{\code{dataList}}{Optional argument; as an alternative to passing in \code{data}, \code{curve_params}, and \code{noise_params} individually, you may create a list containing these three elements (with these names) and pass that in instead. This option may be useful for parallel processing across strata.}
    \item{\code{hessian}}{if \code{TRUE}, the hessian of \code{f}
    at the minimum is returned.}
  }}
}
\value{
A set of lambda estimates for each strata.
}
\description{
Function to estimate seroincidences based on cross-section serology data and longitudinal
response model.
}
