% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.incidence.by.R
\name{est.incidence.by}
\alias{est.incidence.by}
\title{Estimate Seroincidence}
\usage{
est.incidence.by(
  data,
  curve_params,
  noise_params,
  strata = "",
  curve_strata_varnames = strata,
  noise_strata_varnames = strata,
  antigen_isos = unique(pull(data, "antigen_iso")),
  lambda.start = 1/365.25,
  build_graph = TRUE,
  numCores = 1L,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame with cross-sectional serology data per antibody and age, and additional columns to identify possible \code{strata}.}

\item{curve_params}{List of data frames of all longitudinal parameters. Each data frame contains
Monte Carlo samples for each antibody type.}

\item{noise_params}{a \code{\link[=list]{list()}} (or \code{\link[=data.frame]{data.frame()}}, or \code{\link[=tibble]{tibble()}}) containing noise parameters}

\item{strata}{Character vector of stratum-defining variables. Values must be variable names in \code{data}. Default = "".}

\item{curve_strata_varnames}{A subset of \code{strata}. Values must be variable names in \code{curve_params}. Default = "".}

\item{noise_strata_varnames}{A subset of \code{strata}. Values must be variable names in \code{noise_params}. Default = "".}

\item{antigen_isos}{Character vector with one or more antibody names. Values must match \code{data}}

\item{lambda.start}{starting guess for incidence rate, in years/event.}

\item{build_graph}{whether to graph the log-likelihood function across a range of incidence rates (lambda values)}

\item{numCores}{Number of processor cores to use for calculations when computing by strata. If set to more than 1 and package \pkg{parallel} is available, then the computations are executed in parallel. Default = 1L.}

\item{verbose}{logical: if TRUE, print verbose log information to console}

\item{...}{
  Arguments passed on to \code{\link[=.optNll]{.optNll}}, \code{\link[stats:nlm]{stats::nlm}}
  \describe{
    \item{\code{print_graph}}{whether to display the graph as soon as it is created}
    \item{\code{stepmax}}{a positive scalar which gives the maximum allowable
    scaled step length.  \code{stepmax} is used to prevent steps which
    would cause the optimization function to overflow, to prevent the
    algorithm from leaving the area of interest in parameter space, or to
    detect divergence in the algorithm. \code{stepmax} would be chosen
    small enough to prevent the first two of these occurrences, but should
    be larger than any anticipated reasonable step.}
    \item{\code{typsize}}{an estimate of the size of each parameter
    at the minimum.}
    \item{\code{fscale}}{an estimate of the size of \code{f} at the minimum.}
    \item{\code{print.level}}{this argument determines the level of printing
    which is done during the minimization process.  The default
    value of \code{0} means that no printing occurs, a value of \code{1}
    means that initial and final details are printed and a value
    of 2 means that full tracing information is printed.}
    \item{\code{ndigit}}{the number of significant digits in the function \code{f}.}
    \item{\code{gradtol}}{a positive scalar giving the tolerance at which the
    scaled gradient is considered close enough to zero to
    terminate the algorithm.  The scaled gradient is a
    measure of the relative change in \code{f} in each direction
    \code{p[i]} divided by the relative change in \code{p[i]}.}
    \item{\code{steptol}}{A positive scalar providing the minimum allowable
    relative step length.}
    \item{\code{iterlim}}{a positive integer specifying the maximum number of
    iterations to be performed before the program is terminated.}
    \item{\code{check.analyticals}}{a logical scalar specifying whether the
    analytic gradients and Hessians, if they are supplied, should be
    checked against numerical derivatives at the initial parameter
    values. This can help detect incorrectly formulated gradients or
    Hessians.}
  }}
}
\value{
An object of class \code{"seroincidence.ests"}: a list of \verb{"seroincidence.est} objects from \code{\link[=.optNll]{.optNll()}}, one for each stratum, with some meta-data attributes.
}
\description{
Function to estimate seroincidences based on cross-section serology data and longitudinal
response model.
}
