% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_scr.R
\name{est.incidence}
\alias{est.incidence}
\title{Estimate Seroincidence}
\usage{
est.incidence(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{estimate_scr()} was renamed to \code{\link[=estimate_scr]{estimate_scr()}} to create a more
consistent API.
}
\keyword{internal}
