% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.incidence.R
\name{est.incidence}
\alias{est.incidence}
\title{Age specific seroincidence function
This function models seroincidence using maximum likelihood estimation; that is, it finds the value of the seroincidence parameter which maximizes the likelihood (i.e., joint probability) of the data.}
\usage{
est.incidence(
  data,
  curve_params,
  noise_params,
  c.age = NULL,
  antigen_isos = unique(data$antigen_iso),
  start = 0.1,
  iterlim = 100,
  coverage = 0.95,
  verbose = FALSE,
  stepmax = 1,
  ...
)
}
\arguments{
\item{data}{cross-sectional population data}

\item{curve_params}{mcmc samples from distribution of longitudinal decay curve parameters}

\item{noise_params}{a \code{\link[=data.frame]{data.frame()}} containing columns \code{nu}, etc. specifying conditional noise parameters}

\item{c.age}{age category}

\item{antigen_isos}{antigen isotypes: a \code{\link[=character]{character()}} vector of one or more antigen isotype names, which should match the values of the \code{antigen_iso} column in the \code{data} input argument}

\item{start}{starting value for incidence rate}

\item{iterlim}{an \code{\link[=integer]{integer()}}, which provides an upper limit on the number of computational iterations used to search for the maximum likelihood estimate of incidence (passed to \code{\link[stats:nlm]{stats::nlm()}}).}

\item{coverage}{desired confidence interval coverage probability}

\item{verbose}{logical: if TRUE, print verbose log information to console}

\item{stepmax}{a \code{\link[=numeric]{numeric()}}, which limits how aggressively the \code{\link[stats:nlm]{stats::nlm()}} algorithm searches for the maximum likelihood estimate of incidence. If this function output an infinite standard error estimate, consider reducing this parameter.}

\item{...}{
  Arguments passed on to \code{\link[stats:nlm]{stats::nlm}}
  \describe{
    \item{\code{typsize}}{an estimate of the size of each parameter
    at the minimum.}
    \item{\code{fscale}}{an estimate of the size of \code{f} at the minimum.}
    \item{\code{print.level}}{this argument determines the level of printing
    which is done during the minimization process.  The default
    value of \code{0} means that no printing occurs, a value of \code{1}
    means that initial and final details are printed and a value
    of 2 means that full tracing information is printed.}
    \item{\code{ndigit}}{the number of significant digits in the function \code{f}.}
    \item{\code{gradtol}}{a positive scalar giving the tolerance at which the
    scaled gradient is considered close enough to zero to
    terminate the algorithm.  The scaled gradient is a
    measure of the relative change in \code{f} in each direction
    \code{p[i]} divided by the relative change in \code{p[i]}.}
    \item{\code{steptol}}{A positive scalar providing the minimum allowable
    relative step length.}
    \item{\code{check.analyticals}}{a logical scalar specifying whether the
    analytic gradients and Hessians, if they are supplied, should be
    checked against numerical derivatives at the initial parameter
    values. This can help detect incorrectly formulated gradients or
    Hessians.}
  }}
}
\value{
A \code{\link[=data.frame]{data.frame()}} containing the following:
\itemize{
\item \code{est.start}: the starting guess for incidence rate
\item \code{ageCat}: the age category we are analyzing
\item \code{incidence.rate}: the estimated incidence rate, per person year
\item \code{CI.lwr}: lower limit of confidence interval for incidence rate
\item \code{CI.upr}: upper limit of confidence interval for incidence rate
\item \code{coverage}: coverage probability
\item \code{neg.llik}: negative log-likelihood
\item \code{iterations}: the number of iterations used
}
}
\description{
Age specific seroincidence function
This function models seroincidence using maximum likelihood estimation; that is, it finds the value of the seroincidence parameter which maximizes the likelihood (i.e., joint probability) of the data.
}
