% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.incidence.R
\name{est.incidence}
\alias{est.incidence}
\title{Find the maximum likelihood estimate of the incidence rate parameter}
\usage{
est.incidence(
  data,
  curve_params,
  noise_params,
  antigen_isos = unique(data$antigen_iso),
  lambda.start = 0.1,
  stepmin = 1e-08,
  stepmax = 3,
  verbose = FALSE,
  build_graph = FALSE,
  print_graph = build_graph & verbose,
  c.age = NULL,
  ...
)
}
\arguments{
\item{data}{Data frame with cross-sectional serology data per antibody and age, and additional columns}

\item{curve_params}{List of data frames of all longitudinal parameters. Each data frame contains Monte Carlo samples for each antibody type.}

\item{noise_params}{a \code{\link[=data.frame]{data.frame()}} (or \code{\link[tibble:tibble]{tibble::tibble()}}) containing the following variables, specifying noise parameters for each antigen isotype:
\itemize{
\item \code{antigen_iso}: antigen isotype whose noise parameters are being specified on each row
\item \code{nu}: biological noise
\item \code{eps}: measurement noise
\item \code{y.low}: lower limit of detection for the current antigen isotype
\item \code{y.high}: upper limit of detection for the current antigen isotype
}}

\item{antigen_isos}{Character vector with one or more antibody names. Values must match \code{data}}

\item{lambda.start}{starting guess for incidence rate, in years/event.}

\item{stepmin}{A positive scalar providing the minimum allowable relative step length.}

\item{stepmax}{a positive scalar which gives the maximum allowable
    scaled step length.  \code{stepmax} is used to prevent steps which
    would cause the optimization function to overflow, to prevent the
    algorithm from leaving the area of interest in parameter space, or to
    detect divergence in the algorithm. \code{stepmax} would be chosen
    small enough to prevent the first two of these occurrences, but should
    be larger than any anticipated reasonable step.}

\item{verbose}{logical: if TRUE, print verbose log information to console}

\item{build_graph}{whether to graph the log-likelihood function across a range of incidence rates (lambda values)}

\item{print_graph}{whether to display the log-likelihood curve graph in the course of running \code{est.incidence()}}

\item{c.age}{age category to subset data by (optional)}

\item{...}{
  Arguments passed on to \code{\link[stats:nlm]{stats::nlm}}
  \describe{
    \item{\code{typsize}}{an estimate of the size of each parameter
    at the minimum.}
    \item{\code{fscale}}{an estimate of the size of \code{f} at the minimum.}
    \item{\code{ndigit}}{the number of significant digits in the function \code{f}.}
    \item{\code{gradtol}}{a positive scalar giving the tolerance at which the
    scaled gradient is considered close enough to zero to
    terminate the algorithm.  The scaled gradient is a
    measure of the relative change in \code{f} in each direction
    \code{p[i]} divided by the relative change in \code{p[i]}.}
    \item{\code{iterlim}}{a positive integer specifying the maximum number of
    iterations to be performed before the program is terminated.}
    \item{\code{check.analyticals}}{a logical scalar specifying whether the
    analytic gradients and Hessians, if they are supplied, should be
    checked against numerical derivatives at the initial parameter
    values. This can help detect incorrectly formulated gradients or
    Hessians.}
  }}
}
\value{
a \code{"seroincidence"} object, which is a \code{\link[stats:nlm]{stats::nlm()}} fit object with extra meta-data attributes \code{lambda.start}, \code{antigen_isos}, and \code{ll_graph}
}
\description{
This function models seroincidence using maximum likelihood estimation; that is, it finds the value of the seroincidence parameter which maximizes the likelihood (i.e., joint probability) of the data.
}
