% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serocalc.r
\name{fdev}
\alias{fdev}
\title{Calculate negative log-likelihood (deviance)}
\usage{
fdev(log.lambda = log(lambda), lambda = exp(log.lambda), csdata, lnpars, cond)
}
\arguments{
\item{log.lambda}{natural logarithm of incidence parameter, in log(events per person-year).
\itemize{
\item Value of -6 corresponds roughly to 1 day (\code{log(1/365.25)})
\item Value of -4 corresponds roughly to 1 week (\code{log(7 / 365.25)}).
}}

\item{lambda}{incidence parameter, in events per person-year}

\item{csdata}{cross-sectional sample data containing variables \code{y} and \code{a}}

\item{lnpars}{longitudinal antibody decay model parameters \code{alpha}, \code{y1}, and \code{d}}

\item{cond}{measurement noise parameters \code{nu}, \code{eps}, \code{y.low}, and \code{y.high}}
}
\description{
more description to be added here
}
