% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serocalc.r
\name{fdev}
\alias{fdev}
\title{Calculate negative log-likelihood (deviance)}
\usage{
fdev(log.lambda, csdata, lnpars, cond)
}
\arguments{
\item{log.lambda}{Initial guess of incidence rate}

\item{csdata}{cross-sectional sample data containing variables \code{y} and \code{a}}

\item{lnpars}{longitudinal antibody decay model parameters \code{alpha}, \code{y1}, and \code{d}}

\item{cond}{measurement noise parameters \code{nu}, \code{eps}, \code{y.low}, and \code{y.high}}
}
\description{
more description to be added here
}
