% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serocalc.r
\name{fdev}
\alias{fdev}
\title{Calculate negative log-likelihood (deviance)}
\usage{
fdev(log.lambda, csdata, lnpars, cond)
}
\arguments{
\item{log.lambda}{Initial guess of incidence rate}

\item{csdata}{cross-sectional sample data}

\item{lnpars}{longitudinal antibody decay model parameters}

\item{cond}{measurement noise parameters}
}
\description{
more description to be added here
}
