% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdev.r
\name{fdev}
\alias{fdev}
\alias{f_dev}
\title{Calculate negative log-likelihood (deviance)}
\usage{
fdev(lambda, csdata, lnpars, cond)

f_dev(lambda, csdata, lnpars, cond)
}
\arguments{
\item{lambda}{incidence parameter, in events per person-year}

\item{csdata}{cross-sectional sample data containing variables \code{value} and \code{age}}

\item{lnpars}{longitudinal antibody decay model parameters \code{alpha}, \code{y1}, and \code{d}}

\item{cond}{measurement noise parameters \code{nu}, \code{eps}, \code{y.low}, and \code{y.high}}
}
\description{
more description to be added here
}
