% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdditionalData.R
\name{getAdditionalData}
\alias{getAdditionalData}
\title{Get Additional Data}
\usage{
getAdditionalData(fileURL, savePath = NULL)
}
\arguments{
\item{fileURL}{Name of the file URL.}

\item{savePath}{Folder directory and filename to save the downloaded and unzipped (if needed) file. File is saved only
if this argument is not \code{NULL}. Optional. Default = \code{NULL}.}
}
\value{
Data object
}
\description{
Retrieves additional data from internet. The data format must be .RDS or a zipped .RDS. The purpose of this
function is to download data such as longitudinal response parameters from an online repository or population data.
}
\details{
Data for this package is available at: https://osf.io/ne8pc/files/osfstorage

you can save the data into your chosen directory using the optional savePath argument. specify the file path and the file name

***explain this option ti incrtease time to download:
options(timeout = max(300, getOption("timeout")))
}
