% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.curve.params.R
\name{graph.curve.params}
\alias{graph.curve.params}
\title{Graph estimated antibody decay curve}
\usage{
graph.curve.params(
  curve_params,
  antigen_isos = unique(curve_params$antigen_iso),
  verbose = FALSE,
  show_all_curves = FALSE,
  alpha_samples = 0.3,
  quantiles = c(0.1, 0.5, 0.9)
)
}
\arguments{
\item{curve_params}{a \code{\link[=data.frame]{data.frame()}} containing MCMC samples of antibody decay curve parameters}

\item{antigen_isos}{antigen isotypes}

\item{verbose}{verbose output}

\item{show_all_curves}{whether to show individual curves under quantiles}

\item{alpha_samples}{\code{alpha} parameter passed to \link[ggplot2:geom_path]{ggplot2::geom_line}
(has no effect if \code{show_all_curves = FALSE})}

\item{quantiles}{Optional \link{numeric} \link{vector} of quantiles to plot
(e.g., 10\%, 50\%, and 90\% = \code{c(0.1, 0.5, 0.9)}). If \code{NULL}, no quantile
lines are shown.}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object object showing the antibody dynamic
kinetics of selected antigen/isotype combinations, with optional posterior
distribution quantile curves.
}
\description{
Graph estimated antibody decay curve
}
\examples{
curve <-
  typhoid_curves_nostrat_100 |>
  dplyr::filter(antigen_iso \%in\% c("HlyE_IgA", "HlyE_IgG"))

plot1 <- graph.curve.params(curve)

print(plot1)

plot2 <- graph.curve.params(curve, show_all_curves = TRUE,
quantiles = c(0.1, 0.5, 0.9))
print(plot2)

}
