% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.curve.params.R
\name{graph.curve.params}
\alias{graph.curve.params}
\title{Graph estimated antibody decay curves}
\usage{
graph.curve.params(
  curve_params,
  antigen_isos = unique(curve_params$antigen_iso),
  verbose = FALSE,
  show_quantiles = TRUE,
  show_all_curves = TRUE,
  alpha_samples = 0.3,
  ...
)
}
\arguments{
\item{curve_params}{a \code{\link[=data.frame]{data.frame()}} containing MCMC samples of antibody decay curve parameters}

\item{antigen_isos}{antigen isotypes to analyze
(can subset \code{curve_params})}

\item{verbose}{verbose output}

\item{show_quantiles}{whether to show point-wise (over time) quantiles}

\item{show_all_curves}{whether to show individual curves under quantiles}

\item{alpha_samples}{\code{alpha} parameter passed to \link[ggplot2:geom_path]{ggplot2::geom_line}
(has no effect if \code{show_all_curves = FALSE})}

\item{...}{not currently used}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Graph estimated antibody decay curves
}
\examples{
curve <-
  typhoid_curves_nostrat_100 |>
  dplyr::filter(antigen_iso \%in\% c("HlyE_IgA", "HlyE_IgG"))

plot1 <- graph.curve.params(curve)

print(plot1)

plot2 <- graph.curve.params(curve, show_all_curves = TRUE)
show(plot2)

}
