% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_loglik.R
\name{graph_loglik}
\alias{graph_loglik}
\title{Graph log-likelihood of data}
\usage{
graph_loglik(
  ...,
  x = 10^seq(-3, 1, by = 0.1),
  highlight_points = NULL,
  highlight_point_names = NULL,
  log_x = FALSE
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=llik]{llik}}
  \describe{
    \item{\code{data}}{Data frame with cross-sectional serology data per antibody and age, and additional columns}
    \item{\code{antigen_isos}}{Character vector with one or more antibody names. Values must match \code{data}.}
    \item{\code{curve_params}}{List of data frames of all longitudinal parameters. Each data frame contains
Monte Carlo samples for each antibody type.}
    \item{\code{noise_params}}{a \code{\link[=list]{list()}} (or \code{\link[=data.frame]{data.frame()}}, or \code{\link[=tibble]{tibble()}}) containing noise parameters}
    \item{\code{verbose}}{logical: if TRUE, print verbose log information to console}
    \item{\code{lambda}}{incidence parameter, in events per person-year}
  }}

\item{x}{sequence of lambda values to graph}

\item{highlight_points}{a possible highlighted value}

\item{highlight_point_names}{labels for highlighted points}

\item{log_x}{should the x-axis be on a logarithmic scale (\code{TRUE}) or linear scale (\code{FALSE}, default)?}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
\description{
Graph log-likelihood of data
}
