% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estIncidence.R
\name{incidence.age}
\alias{incidence.age}
\title{seroincidence function - Age specific
add some details here}
\usage{
incidence.age(dpop, c.age, start = 0.1)
}
\arguments{
\item{dpop}{cross sectional population data}

\item{c.age}{age category}

\item{start}{starting value for liklihood}
}
\value{

}
\description{
seroincidence function - Age specific
add some details here
}
\examples{


}
