% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{llik}
\alias{llik}
\title{Calculate log-likelihood}
\usage{
llik(...)
}
\description{
Calculates the log-likelihood of a set of cross-sectional antibody response
data, for a given incidence rate (\code{lambda}) value.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{llik()} was renamed to \code{\link[=log_likelihood]{log_likelihood()}} to create a more
consistent API.
}
\keyword{internal}
