% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{llik}
\alias{llik}
\title{Calculate  log-likelihood}
\usage{
llik(
  lambda,
  data,
  antigen_isos,
  curve_params,
  noise_params,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{lambda}{incidence parameter, in events per person-year}

\item{data}{Data frame with cross-sectional serology data per antibody and age, and additional columns}

\item{antigen_isos}{Character vector with one or more antibody names. Values must match \code{data}.}

\item{curve_params}{List of data frames of all longitudinal parameters. Each data frame contains
Monte Carlo samples for each antibody type.}

\item{noise_params}{a \code{\link[=list]{list()}} (or \code{\link[=data.frame]{data.frame()}}, or \code{\link[=tibble]{tibble()}}) containing noise parameters}

\item{verbose}{logical: if TRUE, print verbose log information to console}

\item{...}{additional arguments passed to other functions (not currently used).}
}
\value{
the log-likelihood of the data with the current parameter values
}
\description{
Calculate  log-likelihood
}
