% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.seroincidence.ests.R
\name{plot.seroincidence.ests}
\alias{plot.seroincidence.ests}
\title{Plot \code{seroincidence.ests} log-likelihoods}
\usage{
\method{plot}{seroincidence.ests}(x, ...)
}
\arguments{
\item{x}{a '"seroincidence.ests"' object (from \code{\link[=est.incidence.by]{est.incidence.by()}})}

\item{...}{unused}
}
\value{
aan object of class \code{"ggarrange"}, which is a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} or a \code{\link[=list]{list()}} of \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}s.
}
\description{
Plot \code{seroincidence.ests} log-likelihoods
}
