% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.seroincidence.est.R
\name{plot.seroincidence}
\alias{plot.seroincidence}
\title{Plot the log-likelihood curve for the incidence rate estimate}
\usage{
\method{plot}{seroincidence}(x, log_x = FALSE, ...)
}
\arguments{
\item{x}{a \code{seroincidence} object (from \code{\link[=est.incidence]{est.incidence()}})}

\item{log_x}{should the x-axis be on a logarithmic scale (\code{TRUE}) or linear scale (\code{FALSE}, default)?}

\item{...}{unused}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
\description{
Plot the log-likelihood curve for the incidence rate estimate
}
