% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.decay.curves.R
\name{plot_curve_params_one_ab}
\alias{plot_curve_params_one_ab}
\title{Graph an antibody decay curve model}
\usage{
plot_curve_params_one_ab(
  object,
  verbose = FALSE,
  alpha = 0.4,
  n_curves = 100,
  log_x = FALSE,
  rows_to_graph = 1:min(n_curves, nrow(object)),
  xlim = c(10^-1, 10^3.1),
  ...
)
}
\arguments{
\item{object}{a \code{\link[=data.frame]{data.frame()}} of curve parameters (one or more MCMC samples)}

\item{verbose}{verbose output}

\item{alpha}{(passed to \code{\link[ggplot2:geom_function]{ggplot2::geom_function()}}) how transparent the curves should be:
\itemize{
\item 0 = fully transparent (invisible)
\item 1 = fully opaque
}}

\item{n_curves}{how many curves to plot (see details).}

\item{log_x}{should the x-axis be on a logarithmic scale (\code{TRUE}) or linear scale (\code{FALSE}, default)?}

\item{rows_to_graph}{which rows of \code{curve_params} to plot (overrides \code{n_curves}).}

\item{xlim}{range of x values to graph}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Graph an antibody decay curve model
}
\details{
\subsection{\code{n_curves} and \code{rows_to_graph}}{

In most cases, \code{curve_params} will contain too many rows of MCMC samples for all of these samples to be plotted at once.
\itemize{
\item Setting the  \code{n_curves} argument to a value smaller than the number of rows in \code{curve_params} will cause this function to select the first \code{n_curves} rows to graph.
\item Setting \code{n_curves} larger than the number of rows in ` will result all curves being plotted.
\item If the user directly specifies the \code{rows_to_graph} argument, then \code{n_curves} has no effect.
}
}
}
\examples{
library(dplyr) # loads the `\%>\%` operator and `dplyr::filter()`
"https://osf.io/download/rtw5k/" \%>\%
load_curve_params() \%>\%
dplyr::filter(antigen_iso == "HlyE_IgG") \%>\%
serocalculator:::plot_curve_params_one_ab()
}
