% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_fit.R
\name{postprocess_fit}
\alias{postprocess_fit}
\title{postprocess a fitted incidence model}
\usage{
postprocess_fit(fit, coverage = 0.95, start = NA)
}
\arguments{
\item{fit}{output from \code{\link[stats:nlm]{stats::nlm()}}}

\item{coverage}{desired confidence interval coverage probability}

\item{start}{starting value for incidence rate}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}}; see \code{\link[stats:nlm]{stats::nlm()}} for details on \code{code} variable
}
\description{
postprocess a fitted incidence model
}
