% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.seroincidence.est.R
\name{print.seroincidence}
\alias{print.seroincidence}
\title{Print Method for \code{seroincidence} Object}
\usage{
\method{print}{seroincidence}(x, ...)
}
\arguments{
\item{x}{A list containing output of function \code{\link[=est.incidence.by]{est.incidence.by()}}.}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
Custom \code{\link[=print]{print()}} function to show output of the seroincidence calculator \code{\link[=est.incidence]{est.incidence()}}.
}
\examples{

\dontrun{
# estimate seroincidence
seroincidence <- est.incidence.by(...)

# print the seroincidence object to the console
print(seroincidence)

# or simply type (appropriate print method will be invoked automatically)
seroincidence
}

}
