% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.seroincidence.by.R
\name{print.summary.seroincidence.by}
\alias{print.summary.seroincidence.by}
\title{Print Method for Seroincidence Summary Object}
\usage{
\method{print}{summary.seroincidence.by}(x, ...)
}
\arguments{
\item{x}{A "summary.seroincidence.by" object (constructed by \code{\link[=summary.seroincidence.by]{summary.seroincidence.by()}})}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
Custom \code{\link[=print]{print()}} function for "summary.seroincidence.by" objects (constructed by \code{\link[=summary.seroincidence.by]{summary.seroincidence.by()}})
}
\examples{
\dontrun{
# Estimate seroincidence
seroincidence <- est.incidence.by(...)

# Calculate summary statistics for the seroincidence object
seroincidenceSummary <- summary(seroincidence)

# Print the summary of seroincidence object to the console
print(seroincidenceSummary)

# Or simply type (appropriate print method will be invoked automatically)
seroincidenceSummary
}

}
