% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.seroincidence.R
\name{print.summary.seroincidence.ests}
\alias{print.summary.seroincidence.ests}
\title{Print Method for Seroincidence Summary Object}
\usage{
\method{print}{summary.seroincidence.ests}(x, ...)
}
\arguments{
\item{x}{A "summary.seroincidence.ests" object (constructed by \code{\link[=summary.seroincidence.ests]{summary.seroincidence.ests()}})}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
Custom \code{\link[=print]{print()}} function for "summary.seroincidence.ests" objects (constructed by \code{\link[=summary.seroincidence.ests]{summary.seroincidence.ests()}})
}
\examples{

\dontrun{
# estimate seroincidence
seroincidence <- est.incidence.by(...)

# calculate summary statistics for the seroincidence object
seroincidenceSummary <- summary(seroincidence)

# print the summary of seroincidence object to the console
print(seroincidenceSummary)

# or simply type (appropriate print method will be invoked automatically)
seroincidenceSummary
}

}
