% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sees_pop_data_pakistan_100.R
\docType{data}
\name{sees_pop_data_pk_100}
\alias{sees_pop_data_pk_100}
\title{Small example cross-sectional data set}
\format{
\subsection{\code{sees_pop_data_pk_100}}{

A \code{pop_data} object (from \code{\link[=as_pop_data]{as_pop_data()}}) with 200 rows and 8 columns:
\describe{
\item{id}{Observation ID}
\item{Country}{Country where the participant was living}
\item{cluster}{survey sampling cluster}
\item{catchment}{survey catchment area}
\item{age}{participant's age when sampled, in years}
\item{antigen_iso}{which antigen and isotype are being measured
(data is in long format)}
\item{value}{concentration of antigen isotype, in ELISA units}
}
}
}
\source{
\url{https://osf.io/n6cp3}
}
\usage{
sees_pop_data_pk_100
}
\description{
A subset of data from the SEES data, for examples and testing,
data from Pakistan only.
}
\keyword{datasets}
