% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.cs.multi.R
\name{sim.cs.multi}
\alias{sim.cs.multi}
\title{Simulate multiple data sets}
\usage{
sim.cs.multi(
  nclus = 10,
  lambdas = c(0.05, 0.1, 0.15, 0.2, 0.3),
  num_cores = max(1, parallel::detectCores() - 1),
  rng_seed = 1234,
  renew_params = TRUE,
  add_noise = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{nclus}{number of clusters}

\item{lambdas}{#incidence rate, in events/person*year}

\item{num_cores}{number of cores to use for parallel computations}

\item{rng_seed}{starting seed for random number generator, passed to \code{\link[rngtools:RNGseq]{rngtools::RNGseq()}}}

\item{verbose}{whether to report verbose information}

\item{...}{
  Arguments passed on to \code{\link[=sim.cs]{sim.cs}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Simulate multiple data sets
}
