% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.cs-utils.r
\name{simcs.tinf}
\alias{simcs.tinf}
\title{collect cross-sectional data (age, y(t) set)}
\usage{
simcs.tinf(
  lambda,
  n.smpl,
  age.rng,
  age.fx = NA,
  ablist,
  n.mc = 0,
  renew.params = FALSE,
  ...
)
}
\arguments{
\item{lambda}{seroconversion rate (in events/person-day)}

\item{n.smpl}{number of samples n.smpl (= nr of simulated records)}

\item{age.rng}{age range for output}

\item{age.fx}{age.fx for parameter sample (age.fx = NA for age at infection)}

\item{ablist}{antibody set; subset of 1:7}

\item{n.mc}{\itemize{
\item when \code{n.mc} is in 1:4000 a fixed posterior sample is used
\item when n.mc = 0 a random smaple is chosen
}}

\item{renew.params}{\itemize{
\item \code{renew.params = TRUE} generates a new parameter set for each infection
\item \code{renew.params = FALSE} keeps the one selected at birth, but updates baseline y0
}}

\item{...}{arguments passed to \code{\link[=simresp.tinf]{simresp.tinf()}}}
}
\value{
an \code{\link[=array]{array()}}
}
\description{
collect cross-sectional data (age, y(t) set)
}
