% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_ests_barplot.R
\name{strat_ests_barplot}
\alias{strat_ests_barplot}
\title{Barplot method for \code{summary.seroincidence.by} objects}
\usage{
strat_ests_barplot(
  object,
  xvar,
  color_var,
  alpha = 0.7,
  CIs = FALSE,
  title = "Seroconversion Rate by Group",
  xlab = "Seroconversion rate per 1000 person-years",
  ylab = "Stratification Variable",
  fill_lab,
  fill_palette = NULL,
  ...
)
}
\arguments{
\item{object}{a \code{summary.seroincidence.by} object
(generated by applying the \code{summary()}
method to the output of \code{\link[=estimate_scr_by]{estimate_scr_by()}}).}

\item{xvar}{the name of a stratifying variable in \code{object}}

\item{color_var}{\link{character} which variable in \code{object} to use
to determine bar color}

\item{alpha}{transparency for the points in the graph
(1 = no transparency, 0 = fully transparent)}

\item{CIs}{\link{logical}, if \code{TRUE}, add CI error bars}

\item{...}{unused}

\item{shape}{shape argument for \code{geom_point()}}

\item{dodge_width}{width for jitter}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Barplot method for \code{summary.seroincidence.by} objects
}
\keyword{internal}
