% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.seroincidence.ests.R
\name{strata}
\alias{strata}
\title{Extract strata from an object}
\usage{
strata(x)
}
\arguments{
\item{x}{an object}
}
\value{
the strata of \code{x}
}
\description{
Generic method for extracting strata from objects. See \code{\link[=strata.seroincidence.by]{strata.seroincidence.by()}}
}
