% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.seroincidence.ests.R
\name{strata.seroincidence.by}
\alias{strata.seroincidence.by}
\title{Extract information about strata from a \code{seroincidence.by} object}
\usage{
\method{strata}{seroincidence.by}(x)
}
\arguments{
\item{x}{a \code{seroincidence.by} object (from \code{\link[=est.incidence.by]{est.incidence.by()}})}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} with strata in rows
}
\description{
Extract information about strata from a \code{seroincidence.by} object
}
