% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pop_data.R
\name{summary.pop_data}
\alias{summary.pop_data}
\title{Summarize a cross-sectional antibody survey data set}
\usage{
\method{summary}{pop_data}(object, ...)
}
\arguments{
\item{object}{a \code{pop_data} object}

\item{...}{unused}
}
\value{
a list containing two summary tables: one of \code{age} and one of \code{value}, stratified by \code{antigen_iso}
}
\description{
This function is a \code{summary()} method for \code{pop_data} objects
}
\examples{
library(dplyr)

xs_data <- load_pop_data("https://osf.io/download//n6cp3/") \%>\%
  clean_pop_data()

summary(xs_data)

}
