% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.seroincidence.R
\name{summary.seroincidence.ests}
\alias{summary.seroincidence.ests}
\title{Summary Method for Seroincidence Object}
\usage{
\method{summary}{seroincidence.ests}(
  object,
  ...,
  confidence_level = 0.95,
  showDeviance = TRUE,
  showConvergence = TRUE
)
}
\arguments{
\item{object}{A dataframe containing output of function \code{\link[=est.incidence.by]{est.incidence.by()}}.}

\item{...}{Additional arguments affecting the summary produced.}

\item{confidence_level}{desired confidence interval coverage probability}

\item{showDeviance}{Logical flag (\code{FALSE}/\code{TRUE}) for reporting deviance
(-2*log(likelihood) at estimated seroincidence. Default = \code{TRUE}.}

\item{showConvergence}{Logical flag (\code{FALSE}/\code{TRUE}) for reporting convergence (see
help for \code{\link[=optim]{optim()}} for details). Default = \code{TRUE}.}
}
\value{
A list with the following items:
\describe{
\item{\code{Results}}{Dataframe with maximum likelihood estimate of \code{lambda} (the
seroincidence) (column \code{Lambda}) and corresponding lower (\code{Lambda.lwr}) and upper
(\code{Lambda.upr} bounds.\cr
Optionally \code{Deviance} (Negative log likelihood (NLL) at estimated (maximum likelihood)
\code{lambda}) and \code{Covergence} (Convergence indicator returned by \code{\link[=optim]{optim()}}.
Value of 0 indicates convergence) columns are included.}
\item{\code{Antibodies}}{Character vector with names of input antibodies used in
\code{\link[=est.incidence.by]{est.incidence.by()}}.}
\item{\code{Strata}}{Character with names of strata used in \code{\link[=est.incidence.by]{est.incidence.by()}}.}
\item{\code{CensorLimits}}{List of cutoffs for each of the antibodies used in
\code{\link[=est.incidence.by]{est.incidence.by()}}.}
}
}
\description{
Calculate seroincidence from output of the seroincidence calculator
\code{\link[=est.incidence.by]{est.incidence.by()}}.
}
\examples{

\dontrun{
# estimate seroincidence
seroincidence <- est.incidence.by(...)

# calculate summary statistics for the seroincidence object
seroincidenceSummary <- summary(seroincidence)
}

}
