% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_fit.R
\name{summary.seroincidence}
\alias{summary.seroincidence}
\title{summarize a fitted incidence model}
\usage{
\method{summary}{seroincidence}(object, coverage = 0.95, ...)
}
\arguments{
\item{object}{a \code{\link[=list]{list()}}, outputted by \code{\link[stats:nlm]{stats::nlm()}} or \code{\link[=est.incidence]{est.incidence()}}}

\item{coverage}{desired confidence interval coverage probability}

\item{...}{unused}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}} containing the following:
\itemize{
\item \code{est.start}: the starting guess for incidence rate
\item \code{ageCat}: the age category we are analyzing
\item \code{incidence.rate}: the estimated incidence rate, per person year
\item \code{CI.lwr}: lower limit of confidence interval for incidence rate
\item \code{CI.upr}: upper limit of confidence interval for incidence rate
\item \code{coverage}: coverage probability
\item \code{neg.llik}: negative log-likelihood
\item \code{iterations}: the number of iterations used
\item \code{antigen_isos}: a list of antigen isotypes used in the analysis
\item \code{nlm.exit.code}: information about convergence of the likelihood maximization procedure performed by \code{nlm()} (see "Value" section of \code{\link[stats:nlm]{stats::nlm()}}, component \code{code}); codes 3-5 indicate issues:
\itemize{
\item 1: relative gradient is close to zero, current iterate is probably solution.
\item 2: successive iterates within tolerance, current iterate is probably solution.
\item 3: Last global step failed to locate a point lower than x. Either x is an approximate local minimum of the function, the function is too non-linear for this algorithm, or \code{steptol} is too large.
\item 4: iteration limit exceeded.
\item 5: maximum step size \code{stepmax} exceeded five consecutive times. Either the function is unbounded below, becomes asymptotic to a finite value from above in some direction or \code{stepmax} is too small.
}
}
}
\description{
summarize a fitted incidence model
}
