% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t1.R
\name{t1f}
\alias{t1f}
\title{Calculate time to end of active infection}
\usage{
t1f(
  y0 = 0.74916052,
  b0 = 1,
  mu_b = 0.18432798,
  mu_y = 0.36853621,
  gamma = 0.0013040664
)
}
\arguments{
\item{y0}{initial antibody concentration}

\item{b0}{initial bacteria concentration}

\item{mu_b}{pathogen reproduction rate}

\item{mu_y}{antibody reproduction rate}

\item{gamma}{bacteria destruction rate per antibody}
}
\value{
a \code{\link[=numeric]{numeric()}} vector
}
\description{
Calculate time to end of active infection
}
\examples{
t1f(
  y0 = 0.74916052,
  b0 = 1,
  mu_b = 0.18432798,
  mu_y = 0.36853621,
  gamma = 0.0013040664)
}
\keyword{internal}
