% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y1.R
\name{y1f}
\alias{y1f}
\title{Calculate peak antibody concentration}
\usage{
y1f(y0 = 0.74916052, mu_y = 0.36853621, t1 = t1f(y0 = y0, mu_y = mu_y...))
}
\arguments{
\item{y0}{initial antibody concnetration}

\item{mu_y}{antibody reproduction rate}

\item{...}{
  Arguments passed on to \code{\link[=t1f]{t1f}}
  \describe{
    \item{\code{b0}}{initial bacteria concentration}
    \item{\code{mu_b}}{pathogen reproduction rate}
    \item{\code{gamma}}{bacteria destruction rate per antibody}
  }}
}
\value{
a \code{\link[=numeric]{numeric()}} vector
}
\description{
Calculate peak antibody concentration
}
\examples{
y1f(y0 = 1, mu_y = 1, t1 = 10)
}
